FIX: Fix issues on Landing
