+++ src/imports/components/NavigationBar/index.styl
@@ -9,7 +9,7 @@
     position relative
 
   .bar-content
-    padding-top 40px
+    padding 40px 32px 0
     width 100%
     max-width 1160px
     margin 0 auto
@@ -19,7 +19,7 @@
 
     +iphone()
       display block
-      padding-top 0
+      padding 0
 
     .links
       display flex
