+++ src/imports/icons/index.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 import './index.styl';
 
 export class Logo extends PureComponent {
-  static propsType = {
+  static propTypes = {
     blue: PropTypes.bool
   };
 
@@ -18,227 +18,6 @@ export class Logo extends PureComponent {
     return (
       <svg width="50px" height="51px" viewBox="0 0 50 51">
         <path fill={blue ? '#005EFF' : '#FFF'} d="M40.5206376,45.2214937 C39.4632366,45.2569215 38.407876,45.1055342 37.4030724,44.7742914 C36.5431931,44.5234168 35.7437574,44.0992085 35.0539754,43.5277793 C34.4859771,43.0137537 34.1184488,42.2144439 33.8305945,41.471677 C33.1906326,37.9763029 33.5350298,35.6708981 33.5890025,32.756373 C33.6364466,30.9288302 33.5772331,29.1000759 33.4116637,27.2794303 C33.3962429,27.0481188 33.3011482,25.2875811 32.6817473,25.3312733 C32.5712318,25.2875811 32.1060386,25.5523043 31.8413155,26.1948363 C31.2270549,27.896261 30.7798526,29.2789899 30.1681622,30.9727042 C29.5744626,32.6252965 29.0552967,34.3267212 28.2457064,35.8713681 C27.9501417,36.431656 26.770453,37.0176452 26.2435767,36.8428765 C25.5830538,36.6218455 24.883979,35.6991695 24.7169207,34.9641129 C24.4065001,33.5980788 24.2513094,32.2013629 24.2542977,30.8005056 C24.2542977,30.5177915 24.341682,29.2198769 24.2877093,28.9423031 C24.1489224,29.2378678 23.8919096,29.8110063 23.4909697,30.6257369 C22.6351171,32.3580032 21.9694539,34.0028851 20.645838,35.3573425 C17.913792,38.158782 14.5340737,35.8970694 13.8272885,33.0930598 C13.4443394,31.5741141 13.1693357,30.0268971 12.7401244,28.5208021 C12.4085779,27.3513939 11.7506251,26.2539492 11.568146,25.0742605 C11.4859019,24.5345336 12.3777363,23.5321837 13.1898968,23.4499396 C15.4413289,23.3625553 15.7600247,24.0873313 16.5541943,26.2385285 C17.0065368,27.47476 17.3432236,28.7521136 17.692761,30.0217568 C17.9112218,30.8133563 17.9883257,31.6537881 18.2813203,32.4171161 C18.3661345,32.635577 19.0163769,33.0159559 19.3093715,32.7306717 C20.6500847,31.4549689 21.8440364,30.0334337 22.8689987,28.4925307 C23.5629333,27.4850405 23.9844342,26.349044 24.4573378,25.6216977 C24.8029856,25.112368 25.3713607,24.7990453 25.9865639,24.7786958 C26.8449867,24.8095373 27.5055096,25.6705302 27.634016,26.3207726 C27.975843,28.0427583 27.8910288,29.844418 27.996404,31.6126661 C28.0288498,31.7894358 28.0814524,31.9619035 28.1531818,32.1266916 C28.2221695,31.9947807 28.2839435,31.8592211 28.338231,31.7206114 C29.0527266,29.5719844 29.7492313,27.4156471 30.4765775,25.2721603 C30.967472,23.8251783 32.8539459,22.4861416 34.182702,22.7534349 C35.6425347,23.1286736 36.4958172,24.3983168 37.1846115,25.3595447 C37.759048,26.2090876 38.1406403,27.1740335 38.3026172,28.1866855 C38.55963,29.7287622 38.4182729,31.270839 38.4182729,32.8334768 C38.4182729,33.6276464 38.4722456,34.4243861 38.4876664,35.2185556 C38.5056573,36.331421 38.4285535,37.4545669 38.5442092,38.559722 C38.6341637,39.423285 38.7575298,40.8214346 39.2175827,40.9807825 C39.944929,41.2377953 41.129758,40.8882579 41.8314029,40.387083 C43.1712731,39.4620238 44.3709794,38.3488764 45.3936003,37.0818984 C46.5746501,35.5747638 47.5559975,33.9213316 48.3132656,32.1626734 C49.0462047,30.3921191 49.5372332,28.5308793 49.7730983,26.6291879 C50.0621254,24.5720321 50.0750907,22.4854877 49.8116503,20.424899 C49.5072631,18.5110463 48.8999576,16.6578971 48.0125607,14.9351056 C46.8310171,12.6121146 45.4553936,10.3930619 43.9003559,8.30160533 C42.9134268,6.97284917 41.5512589,5.8728344 40.2147924,4.86020398 C38.9344812,3.908087 37.5630553,3.08505928 36.1205785,2.40316164 C34.5450901,1.63212324 32.9104887,0.897066643 31.2219146,0.480705911 C29.5731607,0.0837028351 27.8749909,-0.068708124 26.1818937,0.028363388 C23.7046736,0.134506108 21.2420071,0.463091017 18.8236173,1.01015227 C17.2070068,1.39567147 15.7394637,2.41087202 14.1973869,3.13307798 C13.7604652,3.33611809 13.2901318,3.474905 12.85321,3.67794511 C12.310913,3.93495791 11.7840368,4.22024211 11.2494501,4.49267568 C10.5323844,4.86020398 9.69195259,5.09665575 9.11881405,5.62353199 C7.76692674,6.87004405 6.34821609,8.12169638 5.3201649,9.62265111 C4.07365284,11.4448718 3.1509769,13.4984041 2.19231916,15.5031039 C1.6648312,16.6669754 1.22831644,17.8699689 0.886694151,19.1012831 C0.526876235,20.3066731 0.290424461,21.5480449 0,22.7714258 C0.280143949,25.1693552 0.398369836,27.5441535 0.873843511,29.844418 C1.2492754,31.5733491 1.86069703,33.242435 2.69092399,34.804765 C3.88563764,37.0770577 5.2355852,39.2642819 6.73116516,41.3508809 C7.69496315,42.7001981 9.06484136,43.7565207 10.2008379,44.9901822 C10.3225641,45.1681225 10.3879373,45.3785706 10.3884573,45.5941622 C10.9024829,45.9694009 11.426789,46.3266487 11.9613756,46.6659056 C12.1567053,46.5091278 12.3340441,46.4088928 12.4754012,46.4474447 C13.3184032,46.7044575 14.017478,47.4215232 14.8527696,47.7325087 C16.1044219,48.2028421 17.3843456,48.5935016 18.6591291,48.9970117 C18.6733534,48.9850095 18.690015,48.9762402 18.7079615,48.9713104 C18.7079615,48.9713104 18.7310927,48.9995818 18.7465134,49.0252831 C19.239978,49.1794908 19.7334426,49.3388387 20.2217669,49.505897 C20.828264,49.7327892 21.4182467,50.0015878 21.9874448,50.3103471 C21.9874448,50.3103471 22.0028656,50.3309081 22.0080058,50.3566094 C22.1416525,50.3746003 22.2650186,50.3951613 22.4089458,50.4105821 C22.4436352,50.3651686 22.4924706,50.3326117 22.5477327,50.3180574 C22.5965651,50.3180574 22.653108,50.3771704 22.7302118,50.449134 C23.5389454,50.543372 24.3588162,50.5999148 25.1898243,50.6187624 L25.1898243,50.5365183 C25.647307,50.3000665 26.0945093,49.9042668 26.5648427,49.8759954 C26.8578373,49.8605747 27.1739631,50.2923562 27.4823784,50.56993 C32.8208309,50.2247754 37.921014,48.2331898 42.0807053,44.8693861 C41.5824405,45.0680832 41.0559436,45.1869136 40.5206376,45.2214937 Z M12.6887218,18.1529059 C13.1307838,17.9704268 13.8015872,17.4306999 14.400427,17.5026635 C15.4490392,17.6260296 17.0527991,17.6285997 16.9474239,19.1167038 C16.890881,19.9005929 15.765165,21.129114 14.9915565,21.2396296 C13.0459696,21.5172034 12.6424595,20.8643909 12.6887218,18.1529059 Z" />
-      </svg>
-    );
-  }
-}
-
-export class Lightning extends PureComponent {
-  render() {
-    return (
-      <svg width="1329px" height="828px" viewBox="0 0 1329 828">
-        <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-          <g transform="translate(-182.000000, -15.000000)" fill="#FFD900" fillRule="nonzero">
-            <g transform="translate(182.000000, 15.000000)">
-              <polygon transform="translate(690.308574, 414.060492) rotate(-5.000000) translate(-690.308574, -414.060492) " points="792.767919 296.675378 884.74549 296.806388 867.057495 316.588924 817.007022 372.268247 783.858411 409.213116 743.372557 454.411625 829.192085 454.411625 727.387851 548.214908 445.297097 808.008079 581.691187 558.695722 490.761794 558.695722 498.885169 548.214908 606.061312 409.213116 632.265748 375.281481 556.403906 375.281481 559.67946 372.268247 619.032508 316.588924 658.470184 279.644055 707.865546 233.266454 935.320051 20.1129044 825.392442 233.266454 801.546405 279.644055" />
-              <polygon transform="translate(933.088048, 730.446869) rotate(-53.000000) translate(-933.088048, -730.446869) " points="945.424411 716.144241 956.498743 716.160204 954.369064 718.57058 948.34286 725.35475 944.351684 729.856247 939.477085 735.363397 949.809973 735.363397 937.552486 746.792728 903.588048 778.446869 920.01024 748.069749 909.062112 748.069749 910.040187 746.792728 922.944465 729.856247 926.099545 725.721894 916.965588 725.721894 917.359973 725.35475 924.506229 718.57058 929.254625 714.069083 935.201951 708.418269 962.588048 682.446869 949.352486 708.418269 946.481363 714.069083" />
-              <polygon transform="translate(1230.386157, 523.962611) rotate(-82.000000) translate(-1230.386157, -523.962611) " points="1258.61343 491.632714 1283.953 491.668796 1279.08001 497.11725 1265.29124 512.452302 1256.15888 522.627559 1245.00514 535.076013 1268.64819 535.076013 1240.6014 560.911064 1162.88616 632.462611 1200.46236 563.797662 1175.41156 563.797662 1177.64953 560.911064 1207.17626 522.627559 1214.39551 513.282199 1193.49578 513.282199 1194.39819 512.452302 1210.74979 497.11725 1221.61477 486.941992 1235.22305 474.168796 1297.88616 415.462611 1267.6014 474.168796 1261.03188 486.941992" />
-              <polygon transform="translate(1132.385053, 195.148416) rotate(-103.000000) translate(-1132.385053, -195.148416) " points="1148.06687 177.121147 1162.14441 177.141266 1159.43719 180.179344 1151.77676 188.730225 1146.70324 194.403987 1140.50671 201.34529 1153.64174 201.34529 1138.06019 215.75101 1094.88505 255.648416 1115.76072 217.360588 1101.84361 217.360588 1103.08693 215.75101 1119.49067 194.403987 1123.50136 189.192979 1111.8904 189.192979 1112.39174 188.730225 1121.47596 180.179344 1127.51206 174.505583 1135.07222 167.383202 1169.88505 134.648416 1153.06019 167.383202 1149.41045 174.505583" />
-              <polygon transform="translate(276.000000, 647.500000) rotate(11.000000) translate(-276.000000, -647.500000) " points="299.836364 620.235617 321.234225 620.266046 317.119251 624.860825 305.475401 637.793149 297.763636 646.374127 288.34492 656.872132 308.31016 656.872132 284.626203 678.659295 219 739 250.731016 681.093615 229.577005 681.093615 231.466845 678.659295 256.400535 646.374127 262.496791 638.493016 244.848128 638.493016 245.61016 637.793149 259.418182 624.860825 268.593048 616.279847 280.084492 605.507981 333 556 307.426203 605.507981 301.87861 616.279847" />
-              <polygon transform="translate(64.424510, 452.141196) rotate(-329.000000) translate(-64.424510, -452.141196) " points="80.3154187 433.815955 94.5806594 433.836408 91.8373439 436.924702 84.074777 445.61692 78.9336005 451.384462 72.6544562 458.440498 85.9646166 458.440498 70.1753118 473.084329 26.4245096 513.641196 47.5785203 474.720511 33.4758465 474.720511 34.7357396 473.084329 51.3581995 451.384462 55.4223706 446.087322 43.6565952 446.087322 44.1646166 445.61692 53.3699642 436.924702 59.4865417 431.157159 67.1475043 423.917053 102.42451 390.641196 85.3753118 423.917053 81.6769161 431.157159" />
-              <polygon transform="translate(176.500000, 144.500000) rotate(-317.000000) translate(-176.500000, -144.500000) " points="204.727273 112.170103 230.066845 112.206186 225.19385 117.654639 211.40508 132.989691 202.272727 143.164948 191.118984 155.613402 214.762032 155.613402 186.715241 181.448454 109 253 146.576203 184.335052 121.525401 184.335052 123.763369 181.448454 153.290107 143.164948 160.509358 133.819588 139.609626 133.819588 140.512032 132.989691 156.863636 117.654639 167.72861 107.479381 181.336898 94.7061856 244 36 213.715241 94.7061856 207.145722 107.479381" />
-            </g>
-          </g>
-        </g>
-      </svg>
-
-    );
-  }
-}
-
-export class ScrollUpBtnIcon extends PureComponent {
-  render() {
-    return (
-      <div className="scroll-to-top-btn-icon">
-        <svg width="14px" height="8px">
-          <g stroke="none" strokeWidth="1" fill="#FFFFFF" fillRule="evenodd">
-            <path
-              d="M7.23076923,0.78359104 C7.47489061,0.776772527 7.72187002,0.86308705 7.91771984,1.04547477 L13.7885455,6.51275841 C14.1938414,6.89019558 14.2266972,7.53613918 13.8619308,7.95551381 C13.4971645,8.37488845 12.8729057,8.40888548 12.4676097,8.03144831 L7.23076923,3.1545717 L1.99392874,8.03144831 C1.5886328,8.40888548 0.964373987,8.37488845 0.599607645,7.95551381 C0.234841304,7.53613918 0.267697031,6.89019558 0.672992966,6.51275841 L6.54381863,1.04547477 C6.73966844,0.86308705 6.98664785,0.776772527 7.23076923,0.78359104 Z"
-            />
-          </g>
-        </svg>
-      </div>
-    );
-  }
-}
-
-export class ApisIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="62px" height="49px" viewBox="0 0 62 49">
-        <g fill="none" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" stroke="#0084FF">
-          <path d="M52.2,28.7 C57.0599743,27.4871863 60.2728874,22.8657731 59.7169604,17.8876999 C59.1610335,12.9096266 55.0077924,9.11082963 50,9 C47.7136247,3.40930845 41.8777497,0.12987513 35.9134657,1.08416056 C29.9491817,2.038446 25.4278693,6.97502932 25,13" />
-          <path d="M54,35 C54,28.372583 48.627417,23 42,23 L40.8,23 C38.0648659,16.2555713 31.1225782,12.2067816 23.905648,13.1470982 C16.6887178,14.0874149 11.0156514,19.7798989 10.1,27 L10,27 C4.4771525,27 4.4408921e-16,31.4771525 0,37 C-4.4408921e-16,42.5228475 4.4771525,47 10,47 L42.1,47 C48.6883314,46.9450953 54.0002288,41.5885602 54,35 Z" />
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class AutomationIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="62px" height="48px" viewBox="0 0 62 48">
-        <g fill="none" stroke="#39C877" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-          <path d="M52.741,20.14 C53.9623053,29.8912495 48.8536697,39.3413102 40.0261325,43.660227 C31.1985954,47.9791438 20.6022818,46.2128083 13.653,39.264 M7.065,25.62 C6.3207602,19.1066924 8.39031651,12.5852041 12.754065,7.69288682 C17.1178134,2.80056951 23.3613105,0.00206701696 29.917,-3.65614318e-15 C36.0172875,-0.00820191744 41.8694582,2.41442834 46.179,6.732" />
-          <path d="M0,16 L7.065,25.619 L16,18 M60,30 L52.741,20.14 L44,28" />
-          <circle cx="30" cy="22.994" r="6" />
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class BusinessAnalysisIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="48px" height="48px" viewBox="0 0 48 48">
-        <g fill="none" stroke="#39C877" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-          <path d="M19.9333333,6.13333333 C8.92445732,6.13333333 4.91458726e-15,15.0577907 3.55271368e-15,26.0666667 C2.1908401e-15,37.0755427 8.92445732,46 19.9333333,46 C30.9422093,46 39.8666667,37.0755427 39.8666667,26.0666667 L19.9333333,26.0666667 L19.9333333,6.13333333 Z" />
-          <path d="M26.0666667,-1.77635684e-15 L26.0666667,19.9333333 L46,19.9333333 C46,14.6466846 43.899887,9.57656331 40.1616618,5.83833816 C36.4234367,2.10011302 31.3533154,-1.43588845e-15 26.0666667,-1.77635684e-15 Z" />
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class DesignThinkingIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="50px" height="47px" viewBox="0 0 50 47">
-        <path fill="#FFC600" d="M35.0363768,22.7809069 C35.3915827,22.3580036 36.0223649,22.3031241 36.4452681,22.65833 C36.8681714,23.013536 36.923051,23.6443181 36.5678451,24.0672214 C34.4206468,26.623646 31.4700576,28.3587883 28.2085328,28.9822173 C27.6660692,29.0859074 27.1422581,28.7302111 27.0385681,28.1877474 C26.9348781,27.6452837 27.2905744,27.1214727 27.833038,27.0177827 C30.6419702,26.4808651 33.1845615,24.9856527 35.0363768,22.7809069 Z M39.473836,18.5517344 C40.9850864,18.9818977 42.4489939,19.6658937 43.8031263,20.6118262 C49.9352508,24.895437 51.8019581,33.1434854 48.1123593,39.6516834 C44.4226334,46.1601056 36.3881354,48.7908308 29.5657029,45.7242332 C29.0619659,45.4978098 28.8371589,44.9058979 29.0635823,44.402161 C29.2900057,43.8984241 29.8819176,43.6736171 30.3856546,43.9000405 C36.266374,46.5433499 43.19187,44.2757437 46.3725007,38.6653305 C49.5532585,33.054693 47.9439767,25.9440796 42.657795,22.2514042 C37.5022117,18.6499588 30.5007369,19.4533699 26.2903015,24.0598606 C27.2162537,25.3529658 27.9418461,26.8180086 28.4116458,28.4200592 C29.7240335,32.8953947 28.7791986,37.534531 26.1982772,41.0646686 C26.2577811,41.4027599 26.1414824,41.7625859 25.8610034,42.0009523 C25.6806382,42.1542366 25.4616527,42.2324164 25.2421254,42.2385825 C23.7949014,43.8369754 21.9710427,45.1330977 19.8439146,45.9766131 C12.881989,48.7373737 4.96308766,45.7410416 1.57254372,39.0631416 C-1.8180005,32.385241 0.435877412,24.2239335 6.7729413,20.2324649 C7.24025406,19.9381229 7.85769726,20.0783434 8.15203929,20.5456561 C8.44638131,21.0129689 8.30616083,21.6304121 7.83884807,21.9247541 C2.37599158,25.3655937 0.433039776,32.4010355 3.35585296,38.157709 C6.27866642,43.914383 13.1051449,46.4973671 19.1066638,44.117457 C20.9184282,43.3989995 22.4749736,42.2996468 23.7152969,40.9444032 C21.0919906,37.2857456 20.3042325,32.5885893 21.6381634,28.241 C21.8001622,27.7130086 22.3595097,27.416313 22.887501,27.5783117 C23.4154923,27.7403105 23.712188,28.2996579 23.5501892,28.8276493 C22.460192,32.3802022 23.0134631,36.2035043 24.999881,39.2784813 C26.90689,36.3211742 27.5503388,32.5902854 26.4924641,28.9828573 C24.8790801,23.4811028 19.7657139,19.8535692 14.1752942,20.0044166 C15.0414657,21.8834145 16.3743627,23.5686074 18.1337206,24.8673565 C18.5780534,25.1953608 18.6723564,25.8214637 18.3443521,26.2657965 C18.0163479,26.7101293 17.390245,26.8044323 16.9459122,26.476428 C14.6493106,24.7810884 12.9804107,22.5190388 12.005439,20.0086854 C11.7557786,19.8605475 11.5723765,19.6054613 11.5261871,19.2959535 C11.4988849,19.1130066 11.5229859,18.9342484 11.5876421,18.7743322 C10.5060489,15.04596 10.8933791,10.9071515 12.9489732,7.31630129 C16.6615633,0.830898858 24.6211246,-1.77996561 31.3976123,1.2689743 C37.9787118,4.23000335 41.3794981,11.6004627 39.473836,18.5517344 Z M37.512783,18.1383352 C39.22919,12.0963834 36.2898144,5.66323217 30.5769903,3.09286561 C24.7416163,0.467360268 17.8870908,2.71575518 14.6846928,8.30991971 C12.9402158,11.3572896 12.586049,14.8619251 13.4579896,18.0382233 C17.8673644,17.7161087 22.0775899,19.4331086 25.0045019,22.5060608 C28.3018679,19.0335953 32.9960142,17.5095764 37.512783,18.1383352 Z" />
-      </svg>
-    );
-  }
-}
-
-export class DesktopAppsIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="62px" height="44px" viewBox="0 0 62 44">
-        <g fill="none" stroke="#0084FF" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-          <path d="M4,37 L4,2 C4,0.8954305 4.8954305,0 6,0 L54,0 C55.1045695,0 56,0.8954305 56,2 L56,37" />
-          <path d="M0,37 L0,40 C0,41.1045695 0.8954305,42 2,42 L58,42 C59.1045695,42 60,41.1045695 60,40 L60,37 L0,37 Z" />
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class ContactUsIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="28px" height="28px" viewBox="0 0 28 28">
-        <path fill="#005EFF" d="M14,0 C6.28064877,0 -2.84217094e-14,6.28064877 -2.84217094e-14,14 C-2.84217094e-14,21.7193512 6.28064877,28 14,28 C21.7193512,28 28,21.7193512 28,14 C28,6.28064877 21.7193512,0 14,0 Z" />
-        <path fill="#FFFFFF" d="M9,8 C9,11.2 9,12.8 9,12.8 C9,12.8 11,13.2 14.9999999,14 L9,15.2 L9,20 L21,14 L9,8 Z" />
-      </svg>
-    );
-  }
-}
-
-export class MobileAppsIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="32px" height="50px" viewBox="0 0 32 50">
-        <rect transform="translate(1, 1)" fill="none" stroke="#0084FF" strokeWidth="2" strokeLinecap="round"
-              strokeLinejoin="round" x="0" y="0" width="30" height="48" rx="4.34782609"
-        />
-        <circle transform="translate(1, 1)" fill="#0084FF" cx="15" cy="41" r="2" />
-      </svg>
-    );
-  }
-}
-
-export class QAIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="50px" height="50px" viewBox="0 0 50 50">
-        <g fill="none" stroke="#39C877" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-          <circle cx="24" cy="24" r="24" />
-          <polyline points="34.4 17.28 22.0923077 31.28 14.4 23.5022222" />
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class PrototypeIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="56px" height="52px" viewBox="0 0 58 54">
-        <g stroke="#FFC600" strokeWidth="1" fill="none" fillRule="evenodd">
-          <rect strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" x="0" y="0" width="56" height="52" rx="6" />
-          <g transform="translate(14.000000, 9.000000)" strokeWidth="2.08">
-            <path d="M7.28359801,30.9284287 C21.4627108,30.9284287 6.86511335,6.02107249 20.4213941,6.02107249" strokeLinecap="square" />
-            <ellipse fill="#101829" cx="3.73333333" cy="31.3970588" rx="3.73333333" ry="3.60294118" />
-            <polygon fill="#101829" transform="translate(22.082821, 5.974843) rotate(-226.000000) translate(-22.082821, -5.974843)"
-                     points="26.3494873 2.61446454 17.816154 1.85719612 18.5067128 10.0924902"
-            />
-          </g>
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class ProjectManagementIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="48px" height="48px" viewBox="0 0 48 48">
-        <g fill="none" stroke="#39C877" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-          <path d="M21.0833333,42.1666667 C9.43932986,42.1666667 1.36187358e-15,32.7273368 0,21.0833333 C-1.70234197e-15,9.43932986 9.43932986,2.04281037e-15 21.0833333,0 C32.7273368,-2.21304456e-15 42.1666667,9.43932986 42.1666667,21.0833333 M21.0833333,33.35 C14.3086404,33.35 8.81666667,27.8580263 8.81666667,21.0833333 C8.81666667,14.3086404 14.3086404,8.81666667 21.0833333,8.81666667 C27.8580263,8.81666667 33.35,14.3086404 33.35,21.0833333" />
-          <path d="M46,46 L27.5233333,27.5233333 M40.7866667,28.75 L27.5233333,27.5233333 L28.5966667,40.8633333" />
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class UIDesignIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="58px" height="51px" viewBox="0 0 58 51">
-        <g stroke="#FFC600" strokeWidth="1" fill="none" fillRule="evenodd" strokeLinecap="round" strokeLinejoin="round">
-          <path
-            d="M6,0 L50,0 C53.3137085,-6.08718376e-16 56,2.6862915 56,6 L56,11.3076923 L0,11.3076923 L0,6 C-4.05812251e-16,2.6862915 2.6862915,6.08718376e-16 6,0 Z M0,11.3076923 L0,43 C4.05812251e-16,46.3137085 2.6862915,49 6,49 L50,49 C53.3137085,49 56,46.3137085 56,43 L56,11.3076923 L0,11.3076923 Z"
-            strokeWidth="2"
-          />
-          <ellipse strokeWidth="1.86666667" cx="5.6" cy="5.65384615" rx="1" ry="1"/>
-          <ellipse strokeWidth="1.86666667" cx="12.1333333" cy="5.65384615" rx="1" ry="1"/>
-          <ellipse strokeWidth="1.86666667" cx="18.6666667" cy="5.65384615" rx="1" ry="1"/>
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class UXDesignIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="54px" height="55px" viewBox="0 0 54 55">
-        <g stroke="#FFC600" strokeWidth="1.79310345" fill="none" fillRule="evenodd" strokeLinecap="round" strokeLinejoin="round">
-          <path
-            d="M0,0 L8.66666667,0 L8.66666667,8.83333333 L0,8.83333333 L0,0 Z M43.3333333,0 L52,0 L52,8.83333333 L43.3333333,8.83333333 L43.3333333,0 Z M0,44.1666667 L8.66666667,44.1666667 L8.66666667,53 L0,53 L0,44.1666667 Z M43.3333333,44.1666667 L52,44.1666667 L52,53 L43.3333333,53 L43.3333333,44.1666667 Z"
-          />
-          <path
-            d="M43.3333333,48.5833333 L8.66666667,48.5833333 M47.6666667,8.83333333 L47.6666667,44.1666667 M8.66666667,4.41666667 L43.3333333,4.41666667 M4.33333333,44.1666667 L4.33333333,8.83333333"
-          />
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class WebAppsIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="56px" height="47px" viewBox="0 0 56 47">
-        <g fill="none" stroke="#0084FF" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-          <path strokeWidth="2" d="M5.54658841,0 L48.4534116,0 C51.5167078,-5.62718381e-16 54,2.48329222 54,5.54658841 L54,12.8571429 L0,12.8571429 L8.8817842e-16,5.54658841 C5.13032832e-16,2.48329222 2.48329222,5.62718381e-16 5.54658841,0 Z M0,12.8571429 L0,39.4534116 C3.75145588e-16,42.5167078 2.48329222,45 5.54658841,45 L48.4534116,45 C51.5167078,45 54,42.5167078 54,39.4534116 L54,12.8571429 L0,12.8571429 Z" />
-          <ellipse strokeWidth="0.924431401" fill="#0084FF" cx="6.47762289" cy="6.5224359" rx="1" ry="1" />
-          <ellipse strokeWidth="0.924431401" fill="#0084FF" cx="12.9486427" cy="6.5224359" rx="1" ry="1" />
-          <ellipse strokeWidth="0.924431401" fill="#0084FF" cx="19.4196625" cy="6.5224359" rx="1" ry="1" />
-        </g>
       </svg>
     );
   }
@@ -258,270 +37,6 @@ export class LPMATitle extends PureComponent {
   }
 }
 
-export class LPMAGraph extends PureComponent {
-  render() {
-    return (
-      <svg viewBox="0 0 1508 391" xmlns="http://www.w3.org/2000/svg">
-        <g fill="none" fillRule="evenodd">
-          <g transform="translate(0 -4)">
-            <g fontFamily="Rubik-Regular, Rubik" fontSize="20.16">
-              <text fill="#8C9BA2" transform="translate(53.76 371.28)">
-                <tspan x="0" y="19">Apr 02, 2018</tspan>
-              </text>
-              <text fill="#8C9BA2" transform="translate(53.76 371.28)">
-                <tspan x="332.64" y="19">Apr 22, 2018</tspan>
-              </text>
-              <text fill="#3D474C" transform="translate(53.76 371.28)">
-                <tspan x="668.64" y="19">Apr 29, 2018</tspan>
-              </text>
-              <text fill="#8C9BA2" transform="translate(53.76 371.28)">
-                <tspan x="1004.64" y="19">Jun 14, 2018</tspan>
-              </text>
-              <text fill="#8C9BA2" transform="translate(53.76 371.28)">
-                <tspan x="1333.92" y="19">Jun 25, 2018</tspan>
-              </text>
-            </g>
-            <path d="M53.76 120.7343c19.7566 9.2015 201.0462 121.1658 411.5932 121.1658 284.838 0 398.9187-183.1001 638.6188-183.1001 239.7 0 309.5805 103.8795 401.308 140.6807" stroke="#6EBF56" strokeWidth="4.86" />
-            <path d="M53.76 120.167c19.7566 8.9545 201.0462 117.9131 411.5932 117.9131 284.838 0 398.9187-176.0394 638.6188-176.0394 239.7 0 304.932 85.6597 401.308 134.7587V330.96H53.76V120.167z" fillOpacity=".04" fill="#49D14E" />
-            <g strokeLinecap="square" strokeWidth="1.68">
-              <path d="M54.6 331.8h1449.84" stroke="#D7D7D7" />
-              <path d="M53.76 170.52h1451.52M53.76 10.92h1451.52" strokeOpacity=".1022" stroke="#738187" />
-            </g>
-            <circle stroke="#6CBF55" strokeWidth="3.36" fill="#FFF" cx="1500.24" cy="198.24" r="5.04" />
-            <circle stroke="#6CBF55" strokeWidth="3.36" fill="#FFF" cx="450.24" cy="241.92" r="5.04" />
-            <circle stroke="#6CBF55" strokeWidth="3.36" fill="#FFF" cx="58.8" cy="122.64" r="5.04" />
-            <g fill="#8C9BA2" fontFamily="Rubik-Regular, Rubik" fontSize="20.16">
-              <text>
-                <tspan x="0" y="339.88">0</tspan>
-              </text>
-              <text>
-                <tspan x="0" y="178.6">5</tspan>
-              </text>
-              <text>
-                <tspan x="0" y="19">10</tspan>
-              </text>
-            </g>
-          </g>
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class BetterYetTitle extends PureComponent {
-  render() {
-    return (
-      <svg width="1432px" height="310px" viewBox="0 0 1432 310">
-        <g transform="translate(-124.000000, -277.000000)" fill="#FFFFFF">
-          <path d="M285,348.220627 C285,365.21549 280.862475,380.670731 272.587302,394.586815 C264.312128,408.502898 253.238165,420.14051 239.365079,429.5 C246.666703,430.977814 252.994682,433.502376 258.349206,437.07376 C263.70373,440.645144 268.145485,445.016945 271.674603,450.189295 C275.203721,455.361645 277.880943,461.211238 279.706349,467.738251 C281.531755,474.265263 282.444444,481.100051 282.444444,488.24282 C282.444444,502.282055 279.219609,514.535413 272.769841,525.003264 C266.320074,535.471114 257.984178,544.091568 247.761905,550.864883 C237.539631,557.638197 226.161438,562.68732 213.626984,566.012402 C201.09253,569.337484 188.61911,571 176.206349,571 C169.3915,571 162.576754,570.507403 155.761905,569.522193 C148.947056,568.536984 142.497385,566.812893 136.412698,564.349869 C134.222211,563.610962 132.396833,560.593804 130.936508,555.298303 C129.476183,550.002802 128.32011,543.537462 127.468254,535.902089 C126.616398,528.266716 125.947092,520.01571 125.460317,511.148825 C124.973543,502.28194 124.608467,493.784636 124.365079,485.656658 C124.121692,477.52868 124,470.386019 124,464.22846 L124,425.251305 C124,416.507572 124.243384,407.825544 124.730159,399.204961 C124.730159,396.98824 124.791005,393.540058 124.912698,388.860313 C125.034392,384.180568 125.156084,378.946722 125.277778,373.158616 C125.399472,367.370511 125.642855,361.274619 126.007937,354.870757 C126.373018,348.466896 126.859785,342.494153 127.468254,336.95235 C128.076723,331.410547 128.928566,326.669297 130.02381,322.72846 C131.119053,318.787622 132.396818,316.324635 133.857143,315.339426 C139.455054,311.64489 145.661342,308.135134 152.47619,304.810052 C159.291039,301.48497 166.34917,298.590961 173.650794,296.127937 C180.952417,293.664914 188.253932,291.694524 195.555556,290.21671 C202.857179,288.738896 209.793618,288 216.365079,288 C225.370415,288 234.071387,289.293068 242.468254,291.879243 C250.865121,294.465418 258.166636,298.344622 264.373016,303.516971 C270.579396,308.689321 275.568764,315.031512 279.34127,322.543734 C283.113775,330.055956 285,338.614834 285,348.220627 Z M242,367.294118 C242,363.071874 240.712277,359.96733 238.136792,357.980392 C235.561308,355.993454 232.1887,355 228.018868,355 C224.584888,355 220.967,355.745091 217.165094,357.235294 C213.363189,358.725498 210.113221,360.712406 207.415094,363.196078 C206.679242,363.94118 206.00472,366.05227 205.391509,369.529412 C204.778299,373.006553 204.287738,376.918279 203.919811,381.264706 C203.551885,385.611133 203.306604,389.709131 203.183962,393.558824 C203.06132,397.408516 203,399.954242 203,401.196078 L203,403.617647 C203,404.7353 203.06132,405.915026 203.183962,407.156863 C203.306604,408.398699 203.490565,409.516335 203.735849,410.509804 C203.981133,411.503273 204.349054,412 204.839623,412 C209.009455,412 213.301865,410.447728 217.716981,407.343137 C222.132098,404.238547 226.117907,400.388912 229.674528,395.794118 C233.23115,391.199323 236.174517,386.294144 238.504717,381.078431 C240.834917,375.862719 242,371.267994 242,367.294118 Z M241,475.65812 C241,471.39599 239.509819,468.387473 236.529412,466.632479 C233.549005,464.877484 230.196097,464 226.470588,464 C222.993447,464 219.330085,464.814807 215.480392,466.444444 C211.6307,468.074082 208.464065,470.017083 205.980392,472.273504 C205.23529,472.774931 204.676472,474.02848 204.303922,476.034188 C203.931371,478.039896 203.683007,480.233606 203.558824,482.615385 C203.43464,484.997163 203.310458,487.253551 203.186275,489.384615 C203.062091,491.51568 203,492.957261 203,493.709402 L203,496.717949 C203,498.22223 203.062091,499.789166 203.186275,501.418803 C203.310458,503.048441 203.558822,504.5527 203.931373,505.931624 C204.303923,507.310548 204.73856,508 205.235294,508 C208.712436,508 212.56207,506.997161 216.784314,504.991453 C221.006557,502.985745 224.918283,500.478648 228.519608,497.470085 C232.120933,494.461523 235.101295,491.014264 237.460784,487.128205 C239.820273,483.242146 241,479.418822 241,475.65812 Z M454,501.68595 C454,510.115745 451.390963,518.359464 446.172811,526.417355 C440.954659,534.475247 434.401729,541.665258 426.513825,547.987603 C418.625921,554.309949 410.192057,559.392543 401.211982,563.235537 C392.231906,567.078532 383.980068,569 376.456221,569 C363.59287,569 352.125242,565.714909 342.052995,559.144628 C331.980748,552.574347 323.546885,544.206662 316.751152,534.041322 C309.955419,523.875982 304.798021,512.657086 301.278802,500.384298 C297.759583,488.111509 296,476.148819 296,464.495868 C296,452.099112 297.698908,438.834781 301.096774,424.702479 C304.494641,410.570177 309.652039,397.553779 316.569124,385.652893 C323.48621,373.752007 332.22345,363.83475 342.781106,355.900826 C353.338762,347.966902 365.898545,344 380.460829,344 C388.712791,344 396.600577,345.425606 404.124424,348.27686 C411.648271,351.128113 418.201201,355.156999 423.78341,360.363636 C429.365619,365.570274 433.85559,371.768559 437.253456,378.958678 C440.651323,386.148796 442.35023,394.206567 442.35023,403.132231 C442.35023,414.785182 439.923219,425.694164 435.069124,435.859504 C430.215029,446.024844 423.78345,454.826409 415.774194,462.264463 C407.764937,469.702517 398.542295,475.714853 388.105991,480.301653 C377.669687,484.888453 366.990838,487.42975 356.069124,487.92562 C357.039943,495.115738 359.891681,500.322298 364.624424,503.545455 C369.357167,506.768611 374.999967,508.380165 381.552995,508.380165 C386.40709,508.380165 391.139762,507.140508 395.751152,504.661157 C400.362542,502.181806 404.852513,499.392577 409.221198,496.293388 C413.589884,493.194199 417.897828,490.404971 422.145161,487.92562 C426.392494,485.446269 430.700439,484.206612 435.069124,484.206612 C439.923219,484.206612 444.291839,485.818166 448.175115,489.041322 C452.058391,492.264479 454,496.479313 454,501.68595 Z M411,410.330435 C411,407.089839 410.132752,403.724655 408.39823,400.234783 C406.663708,396.74491 403.938072,395 400.221239,395 C394.769884,395 390.185859,396.433319 386.469027,399.3 C382.752194,402.166681 379.65488,405.781138 377.176991,410.143478 C374.699103,414.505819 372.778768,419.179685 371.415929,424.165217 C370.053091,429.15075 369.247789,433.762298 369,438 C372.469044,437.25217 376.619445,436.255079 381.451327,435.008696 C386.28321,433.762313 390.867235,432.079721 395.20354,429.96087 C399.539845,427.842018 403.256622,425.162335 406.353982,421.921739 C409.451343,418.681143 411,414.817414 411,410.330435 Z M556.169767,581 C538.88751,581 525.439195,578.712009 515.824419,574.135959 C506.209642,569.559909 499.029093,563.066963 494.282558,554.656925 C489.536023,546.246887 486.675974,536.105523 485.702326,524.232529 C484.728677,512.359534 484.24186,499.126291 484.24186,484.532402 C484.24186,472.906761 484.606973,461.219457 485.337209,449.47014 C486.067446,437.720822 487.162783,425.167792 488.623256,411.810673 C485.215487,411.810673 482.537994,411.872511 480.590698,411.996188 C478.643401,412.119865 477.061246,412.181703 475.844186,412.181703 C468.055,412.181703 461.300416,410.45025 455.580233,406.987294 C449.860049,403.524337 447,398.206304 447,391.033037 C447,386.580664 448.643007,382.623058 451.92907,379.160102 C455.215133,375.697145 459.413928,372.728941 464.525581,370.2554 C469.637235,367.78186 475.174776,365.803057 481.138372,364.318933 C487.101968,362.834808 492.761214,361.845407 498.116279,361.350699 C498.116279,354.424785 498.603096,346.756925 499.576744,338.346887 C500.550392,329.936849 502.315104,322.083476 504.87093,314.786531 C507.426757,307.489587 510.895327,301.367666 515.276744,296.420584 C519.658161,291.473503 525.256555,289 532.072093,289 C535.966686,289 539.435256,290.174914 542.477907,292.524778 C545.520558,294.874641 548.015494,297.842845 549.962791,301.429479 C551.910087,305.016113 553.370538,309.035556 554.344186,313.487929 C555.317834,317.940302 555.804651,322.268933 555.804651,326.473952 C555.804651,331.421033 555.439539,336.491715 554.709302,341.68615 C553.979066,346.880585 552.64032,353.311694 550.693023,360.97967 C554.831028,360.484962 557.630225,360.175774 559.090698,360.052097 C560.55117,359.92842 561.889916,359.866582 563.106977,359.866582 C567.244982,359.866582 571.565481,360.237607 576.068605,360.97967 C580.571728,361.721732 584.709671,362.958484 588.482558,364.689962 C592.255445,366.42144 595.358903,368.709431 597.793023,371.554003 C600.227144,374.398574 601.444186,377.799641 601.444186,381.757306 C601.444186,386.209679 599.679475,390.167285 596.15,393.630241 C592.620525,397.093198 588.178321,400.061402 582.823256,402.534943 C577.46819,405.008483 571.748092,406.925449 565.662791,408.285896 C559.577489,409.646343 553.857391,410.573907 548.502326,411.068615 C547.041853,421.210131 546.007367,430.547607 545.398837,439.081321 C544.790307,447.615036 544.120934,456.210461 543.390698,464.867853 C542.903874,469.072872 542.477909,473.339665 542.112791,477.668361 C541.747673,481.997057 541.565116,486.387525 541.565116,490.839898 C541.565116,494.055501 541.68682,497.827594 541.930233,502.15629 C542.173645,506.484986 543.025574,510.628104 544.486047,514.585769 C545.946519,518.543434 548.258899,521.944501 551.423256,524.789072 C554.587613,527.633644 558.968964,529.055909 564.567442,529.055909 C567.975211,529.055909 571.504633,528.746721 575.155814,528.128335 C578.806995,527.50995 582.458121,527.200762 586.109302,527.200762 C592.438016,527.200762 597.001924,529.426915 599.801163,533.879288 C602.600402,538.331661 604,544.020719 604,550.946633 C604,556.141068 602.478697,560.655212 599.436047,564.489199 C596.393396,568.323187 592.559713,571.415067 587.934884,573.76493 C583.310054,576.114794 578.137625,577.908084 572.417442,579.144854 C566.697258,580.381624 561.281421,581 556.169767,581 Z M711.169767,581 C693.88751,581 680.439195,578.712009 670.824419,574.135959 C661.209642,569.559909 654.029094,563.066963 649.282558,554.656925 C644.536023,546.246887 641.675974,536.105523 640.702326,524.232529 C639.728677,512.359534 639.24186,499.126291 639.24186,484.532402 C639.24186,472.906761 639.606973,461.219457 640.337209,449.47014 C641.067446,437.720822 642.162783,425.167792 643.623256,411.810673 C640.215487,411.810673 637.537994,411.872511 635.590698,411.996188 C633.643401,412.119865 632.061246,412.181703 630.844186,412.181703 C623.055,412.181703 616.300416,410.45025 610.580233,406.987294 C604.860049,403.524337 602,398.206304 602,391.033037 C602,386.580664 603.643007,382.623058 606.92907,379.160102 C610.215133,375.697145 614.413928,372.728941 619.525581,370.2554 C624.637235,367.78186 630.174776,365.803057 636.138372,364.318933 C642.101968,362.834808 647.761214,361.845407 653.116279,361.350699 C653.116279,354.424785 653.603096,346.756925 654.576744,338.346887 C655.550392,329.936849 657.315104,322.083476 659.87093,314.786531 C662.426757,307.489587 665.895327,301.367666 670.276744,296.420584 C674.658161,291.473503 680.256555,289 687.072093,289 C690.966686,289 694.435256,290.174914 697.477907,292.524778 C700.520558,294.874641 703.015494,297.842845 704.962791,301.429479 C706.910087,305.016113 708.370538,309.035556 709.344186,313.487929 C710.317834,317.940302 710.804651,322.268933 710.804651,326.473952 C710.804651,331.421033 710.439539,336.491715 709.709302,341.68615 C708.979066,346.880585 707.64032,353.311694 705.693023,360.97967 C709.831028,360.484962 712.630225,360.175774 714.090698,360.052097 C715.55117,359.92842 716.889916,359.866582 718.106977,359.866582 C722.244982,359.866582 726.565481,360.237607 731.068605,360.97967 C735.571728,361.721732 739.709671,362.958484 743.482558,364.689962 C747.255445,366.42144 750.358903,368.709431 752.793023,371.554003 C755.227144,374.398574 756.444186,377.799641 756.444186,381.757306 C756.444186,386.209679 754.679475,390.167285 751.15,393.630241 C747.620525,397.093198 743.178321,400.061402 737.823256,402.534943 C732.46819,405.008483 726.748092,406.925449 720.662791,408.285896 C714.577489,409.646343 708.857391,410.573907 703.502326,411.068615 C702.041853,421.210131 701.007367,430.547607 700.398837,439.081321 C699.790307,447.615036 699.120934,456.210461 698.390698,464.867853 C697.903874,469.072872 697.477909,473.339665 697.112791,477.668361 C696.747673,481.997057 696.565116,486.387525 696.565116,490.839898 C696.565116,494.055501 696.68682,497.827594 696.930233,502.15629 C697.173645,506.484986 698.025574,510.628104 699.486047,514.585769 C700.946519,518.543434 703.258899,521.944501 706.423256,524.789072 C709.587613,527.633644 713.968964,529.055909 719.567442,529.055909 C722.975211,529.055909 726.504633,528.746721 730.155814,528.128335 C733.806995,527.50995 737.458121,527.200762 741.109302,527.200762 C747.438016,527.200762 752.001924,529.426915 754.801163,533.879288 C757.600402,538.331661 759,544.020719 759,550.946633 C759,556.141068 757.478697,560.655212 754.436047,564.489199 C751.393396,568.323187 747.559713,571.415067 742.934884,573.76493 C738.310054,576.114794 733.137626,577.908084 727.417442,579.144854 C721.697258,580.381624 716.281421,581 711.169767,581 Z M918,501.68595 C918,510.115745 915.390963,518.359464 910.172811,526.417355 C904.954659,534.475247 898.401729,541.665258 890.513825,547.987603 C882.625921,554.309949 874.192057,559.392543 865.211982,563.235537 C856.231906,567.078532 847.980068,569 840.456221,569 C827.59287,569 816.125242,565.714909 806.052995,559.144628 C795.980748,552.574347 787.546885,544.206662 780.751152,534.041322 C773.955419,523.875982 768.798021,512.657086 765.278802,500.384298 C761.759583,488.111509 760,476.148819 760,464.495868 C760,452.099112 761.698908,438.834781 765.096774,424.702479 C768.494641,410.570177 773.652039,397.553779 780.569124,385.652893 C787.48621,373.752007 796.22345,363.83475 806.781106,355.900826 C817.338762,347.966902 829.898545,344 844.460829,344 C852.712791,344 860.600577,345.425606 868.124424,348.27686 C875.648271,351.128113 882.201201,355.156999 887.78341,360.363636 C893.365619,365.570274 897.85559,371.768559 901.253456,378.958678 C904.651323,386.148796 906.35023,394.206567 906.35023,403.132231 C906.35023,414.785182 903.923219,425.694164 899.069124,435.859504 C894.215029,446.024844 887.78345,454.826409 879.774194,462.264463 C871.764937,469.702517 862.542295,475.714853 852.105991,480.301653 C841.669687,484.888453 830.990838,487.42975 820.069124,487.92562 C821.039943,495.115738 823.891681,500.322298 828.624424,503.545455 C833.357167,506.768611 838.999967,508.380165 845.552995,508.380165 C850.40709,508.380165 855.139762,507.140508 859.751152,504.661157 C864.362542,502.181806 868.852513,499.392577 873.221198,496.293388 C877.589884,493.194199 881.897828,490.404971 886.145161,487.92562 C890.392494,485.446269 894.700439,484.206612 899.069124,484.206612 C903.923219,484.206612 908.291839,485.818166 912.175115,489.041322 C916.058391,492.264479 918,496.479313 918,501.68595 Z M862,410.330435 C862,407.089839 861.132752,403.724655 859.39823,400.234783 C857.663708,396.74491 854.938072,395 851.221239,395 C845.769884,395 841.185859,396.433319 837.469027,399.3 C833.752194,402.166681 830.65488,405.781138 828.176991,410.143478 C825.699103,414.505819 823.778768,419.179685 822.415929,424.165217 C821.053091,429.15075 820.247789,433.762298 820,438 C823.469044,437.25217 827.619445,436.255079 832.451327,435.008696 C837.28321,433.762313 841.867235,432.079721 846.20354,429.96087 C850.539845,427.842018 854.256622,425.162335 857.353982,421.921739 C860.451343,418.681143 862,414.817414 862,410.330435 Z M1061.34128,428.442907 C1057.92858,428.690889 1055.00345,428.380916 1052.56581,427.512976 C1050.12817,426.645036 1047.69056,425.715115 1045.25292,424.723183 C1043.05904,423.731252 1040.86519,422.801331 1038.67131,421.933391 C1036.47743,421.065451 1034.03983,420.755477 1031.35842,421.00346 C1026.48313,421.251443 1022.15638,422.615328 1018.37803,425.095156 C1014.59968,427.574984 1011.18704,430.674722 1008.13998,434.394464 C1005.09293,438.114205 1002.4725,442.267855 1000.27862,446.855536 C998.084741,451.443218 996.378417,455.968836 995.159595,460.432526 C994.428302,463.408319 993.87984,467.933937 993.514194,474.009516 C993.148547,480.085094 992.965727,486.594544 992.965727,493.538062 C992.721963,500.729563 992.600082,507.548987 992.600082,513.99654 C992.600082,520.444092 992.356321,525.527663 991.868793,529.247405 C991.625028,533.215129 991.015627,537.864737 990.040569,543.196367 C989.065512,548.527997 987.542008,553.549573 985.470011,558.261246 C983.398014,562.972919 980.655706,566.940584 977.243006,570.16436 C973.830305,573.388136 969.442613,575 964.079797,575 C959.935804,575 956.340334,573.760105 953.29328,571.280277 C950.246225,568.800449 947.686738,565.514726 945.614741,561.42301 C943.542744,557.331294 941.83642,552.557697 940.495716,547.102076 C939.155012,541.646455 938.119029,536.066926 937.387736,530.363322 C936.656443,524.411735 936.107982,518.460237 935.742335,512.508651 C935.376689,506.557064 935.132928,500.91554 935.011046,495.58391 C934.889164,490.25228 934.889164,485.354694 935.011046,480.891003 C935.132928,476.427313 935.193868,472.831617 935.193868,470.103806 C934.950104,461.424409 934.523523,452.683147 933.914112,443.879758 C933.304701,435.076369 933,426.335107 933,417.655709 C933,412.200088 933.243761,406.124601 933.731289,399.429066 C934.218818,392.733531 935.559502,386.472059 937.753381,380.644464 C939.94726,374.816868 943.177089,369.919282 947.442965,365.951557 C951.708841,361.983833 957.620037,360 965.176731,360 C969.320725,360 972.855255,361.363885 975.780427,364.091696 C978.705599,366.819506 981.082266,370.167224 982.910498,374.134948 C984.738731,378.102673 986.140355,382.318317 987.115412,386.782007 C988.090469,391.245697 988.699871,395.337352 988.943635,399.057093 C994.793979,389.385765 1001.37552,381.512429 1008.68845,375.436851 C1016.00138,369.361273 1025.50804,366.07555 1037.20873,365.579585 C1043.05908,365.331602 1048.84839,365.889555 1054.57685,367.25346 C1060.30532,368.617366 1065.54617,370.849177 1070.29958,373.948962 C1075.05298,377.048747 1078.95315,381.078407 1082.00021,386.038062 C1085.04726,390.997718 1086.69264,396.825226 1086.93641,403.520761 C1087.42394,411.45621 1085.10821,417.531697 1079.98916,421.747405 C1074.87011,425.963112 1068.65421,428.194924 1061.34128,428.442907 Z M1257,312.227273 C1257,323.10452 1255.60016,335.094033 1252.80042,348.196172 C1250.00069,361.298311 1246.40979,374.523857 1242.0276,387.873206 C1237.64541,401.222555 1232.89811,414.200894 1227.78556,426.808612 C1222.67301,439.416331 1217.92571,450.540622 1213.54352,460.181818 C1211.10898,465.620442 1208.1267,472.480421 1204.5966,480.761962 C1201.06651,489.043502 1197.11043,497.881133 1192.72824,507.27512 C1188.34605,516.669106 1183.65961,526.124753 1178.66879,535.642344 C1173.67797,545.159936 1168.50463,553.688559 1163.14862,561.228469 C1157.79261,568.768379 1152.31496,574.94854 1146.7155,579.769139 C1141.11604,584.589737 1135.51666,587 1129.9172,587 C1122.12664,587 1116.58813,584.095324 1113.30149,578.285885 C1110.01485,572.476446 1108.37155,565.987277 1108.37155,558.818182 C1108.37155,551.896297 1109.52794,544.541905 1111.84076,536.754785 C1114.15359,528.967664 1117.075,521.242463 1120.6051,513.578947 C1124.13519,505.915432 1128.03041,498.499238 1132.29087,491.330144 C1136.55133,484.161049 1140.50741,477.733681 1144.15924,472.047847 C1139.29014,470.811796 1134.54284,468.586938 1129.9172,465.373206 C1125.29156,462.159474 1121.88324,458.327773 1119.69214,453.87799 C1115.79687,445.225635 1111.77992,435.460979 1107.64119,424.583732 C1103.50246,413.706485 1099.78983,402.643996 1096.50318,391.395933 C1093.21654,380.14787 1090.47772,368.899976 1088.28662,357.651914 C1086.09553,346.403851 1085,336.082982 1085,326.688995 C1085,322.239212 1085.4869,317.356885 1086.46072,312.041866 C1087.43454,306.726847 1089.07784,301.782718 1091.39066,297.20933 C1093.70348,292.635942 1096.74662,288.804242 1100.52017,285.714115 C1104.29372,282.623988 1108.98016,281.078947 1114.57962,281.078947 C1120.90944,281.078947 1126.50882,283.48921 1131.37792,288.309809 C1136.24702,293.130407 1140.56828,299.248767 1144.34183,306.665072 C1148.11538,314.081377 1151.3411,322.362793 1154.01911,331.509569 C1156.69711,340.656346 1159.0099,349.679381 1160.95754,358.578947 C1162.90518,367.478513 1164.48761,375.636327 1165.70488,383.052632 C1166.92216,390.468937 1168.01769,396.278288 1168.99151,400.480861 C1169.96533,396.772709 1171.60862,391.148762 1173.92144,383.608852 C1176.23426,376.068942 1179.03395,367.725724 1182.32059,358.578947 C1185.60724,349.432171 1189.38073,340.038326 1193.64119,330.397129 C1197.90165,320.755933 1202.4055,311.980103 1207.15287,304.069378 C1211.90024,296.158653 1216.95185,289.669483 1222.30786,284.601675 C1227.66386,279.533866 1233.14151,277 1238.74098,277 C1242.3928,277 1245.43594,278.359636 1247.87049,281.078947 C1250.30504,283.798259 1252.19178,287.011943 1253.53079,290.720096 C1254.86979,294.428248 1255.78273,298.259948 1256.26964,302.215311 C1256.75655,306.170674 1257,309.507961 1257,312.227273 Z M1407,501.68595 C1407,510.115745 1404.39096,518.359464 1399.17281,526.417355 C1393.95466,534.475247 1387.40173,541.665258 1379.51382,547.987603 C1371.62592,554.309949 1363.19206,559.392543 1354.21198,563.235537 C1345.23191,567.078532 1336.98007,569 1329.45622,569 C1316.59287,569 1305.12524,565.714909 1295.053,559.144628 C1284.98075,552.574347 1276.54688,544.206662 1269.75115,534.041322 C1262.95542,523.875982 1257.79802,512.657086 1254.2788,500.384298 C1250.75958,488.111509 1249,476.148819 1249,464.495868 C1249,452.099112 1250.69891,438.834781 1254.09677,424.702479 C1257.49464,410.570177 1262.65204,397.553779 1269.56912,385.652893 C1276.48621,373.752007 1285.22345,363.83475 1295.78111,355.900826 C1306.33876,347.966902 1318.89854,344 1333.46083,344 C1341.71279,344 1349.60058,345.425606 1357.12442,348.27686 C1364.64827,351.128113 1371.2012,355.156999 1376.78341,360.363636 C1382.36562,365.570274 1386.85559,371.768559 1390.25346,378.958678 C1393.65132,386.148796 1395.35023,394.206567 1395.35023,403.132231 C1395.35023,414.785182 1392.92322,425.694164 1388.06912,435.859504 C1383.21503,446.024844 1376.78345,454.826409 1368.77419,462.264463 C1360.76494,469.702517 1351.54229,475.714853 1341.10599,480.301653 C1330.66969,484.888453 1319.99084,487.42975 1309.06912,487.92562 C1310.03994,495.115738 1312.89168,500.322298 1317.62442,503.545455 C1322.35717,506.768611 1327.99997,508.380165 1334.553,508.380165 C1339.40709,508.380165 1344.13976,507.140508 1348.75115,504.661157 C1353.36254,502.181806 1357.85251,499.392577 1362.2212,496.293388 C1366.58988,493.194199 1370.89783,490.404971 1375.14516,487.92562 C1379.39249,485.446269 1383.70044,484.206612 1388.06912,484.206612 C1392.92322,484.206612 1397.29184,485.818166 1401.17512,489.041322 C1405.05839,492.264479 1407,496.479313 1407,501.68595 Z M1350,410.330435 C1350,407.089839 1349.13275,403.724655 1347.39823,400.234783 C1345.66371,396.74491 1342.93807,395 1339.22124,395 C1333.76988,395 1329.18586,396.433319 1325.46903,399.3 C1321.75219,402.166681 1318.65488,405.781138 1316.17699,410.143478 C1313.6991,414.505819 1311.77877,419.179685 1310.41593,424.165217 C1309.05309,429.15075 1308.24779,433.762298 1308,438 C1311.46904,437.25217 1315.61944,436.255079 1320.45133,435.008696 C1325.28321,433.762313 1329.86723,432.079721 1334.20354,429.96087 C1338.53984,427.842018 1342.25662,425.162335 1345.35398,421.921739 C1348.45134,418.681143 1350,414.817414 1350,410.330435 Z M1508.16977,581 C1490.88751,581 1477.4392,578.712009 1467.82442,574.135959 C1458.20964,569.559909 1451.02909,563.066963 1446.28256,554.656925 C1441.53602,546.246887 1438.67597,536.105523 1437.70233,524.232529 C1436.72868,512.359534 1436.24186,499.126291 1436.24186,484.532402 C1436.24186,472.906761 1436.60697,461.219457 1437.33721,449.47014 C1438.06745,437.720822 1439.16278,425.167792 1440.62326,411.810673 C1437.21549,411.810673 1434.53799,411.872511 1432.5907,411.996188 C1430.6434,412.119865 1429.06125,412.181703 1427.84419,412.181703 C1420.055,412.181703 1413.30042,410.45025 1407.58023,406.987294 C1401.86005,403.524337 1399,398.206304 1399,391.033037 C1399,386.580664 1400.64301,382.623058 1403.92907,379.160102 C1407.21513,375.697145 1411.41393,372.728941 1416.52558,370.2554 C1421.63723,367.78186 1427.17478,365.803057 1433.13837,364.318933 C1439.10197,362.834808 1444.76121,361.845407 1450.11628,361.350699 C1450.11628,354.424785 1450.6031,346.756925 1451.57674,338.346887 C1452.55039,329.936849 1454.3151,322.083476 1456.87093,314.786531 C1459.42676,307.489587 1462.89533,301.367666 1467.27674,296.420584 C1471.65816,291.473503 1477.25656,289 1484.07209,289 C1487.96669,289 1491.43526,290.174914 1494.47791,292.524778 C1497.52056,294.874641 1500.01549,297.842845 1501.96279,301.429479 C1503.91009,305.016113 1505.37054,309.035556 1506.34419,313.487929 C1507.31783,317.940302 1507.80465,322.268933 1507.80465,326.473952 C1507.80465,331.421033 1507.43954,336.491715 1506.7093,341.68615 C1505.97907,346.880585 1504.64032,353.311694 1502.69302,360.97967 C1506.83103,360.484962 1509.63023,360.175774 1511.0907,360.052097 C1512.55117,359.92842 1513.88992,359.866582 1515.10698,359.866582 C1519.24498,359.866582 1523.56548,360.237607 1528.0686,360.97967 C1532.57173,361.721732 1536.70967,362.958484 1540.48256,364.689962 C1544.25545,366.42144 1547.3589,368.709431 1549.79302,371.554003 C1552.22714,374.398574 1553.44419,377.799641 1553.44419,381.757306 C1553.44419,386.209679 1551.67948,390.167285 1548.15,393.630241 C1544.62052,397.093198 1540.17832,400.061402 1534.82326,402.534943 C1529.46819,405.008483 1523.74809,406.925449 1517.66279,408.285896 C1511.57749,409.646343 1505.85739,410.573907 1500.50233,411.068615 C1499.04185,421.210131 1498.00737,430.547607 1497.39884,439.081321 C1496.79031,447.615036 1496.12093,456.210461 1495.3907,464.867853 C1494.90387,469.072872 1494.47791,473.339665 1494.11279,477.668361 C1493.74767,481.997057 1493.56512,486.387525 1493.56512,490.839898 C1493.56512,494.055501 1493.68682,497.827594 1493.93023,502.15629 C1494.17364,506.484986 1495.02557,510.628104 1496.48605,514.585769 C1497.94652,518.543434 1500.2589,521.944501 1503.42326,524.789072 C1506.58761,527.633644 1510.96896,529.055909 1516.56744,529.055909 C1519.97521,529.055909 1523.50463,528.746721 1527.15581,528.128335 C1530.80699,527.50995 1534.45812,527.200762 1538.1093,527.200762 C1544.43802,527.200762 1549.00192,529.426915 1551.80116,533.879288 C1554.6004,538.331661 1556,544.020719 1556,550.946633 C1556,556.141068 1554.4787,560.655212 1551.43605,564.489199 C1548.3934,568.323187 1544.55971,571.415067 1539.93488,573.76493 C1535.31005,576.114794 1530.13763,577.908084 1524.41744,579.144854 C1518.69726,580.381624 1513.28142,581 1508.16977,581 Z" />
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class BetterYetLogo extends PureComponent {
-  render() {
-    return (
-      <svg width="237px" height="241px" viewBox="0 0 237 241">
-        <defs>
-          <rect x="0" y="0" width="221" height="221" rx="49.8" />
-          <filter x="-8.1%" y="-4.5%" width="116.3%" height="118.1%" filterUnits="objectBoundingBox">
-            <feMorphology radius="2" operator="erode" in="SourceAlpha" result="shadowSpreadOuter1" />
-            <feOffset dx="0" dy="12" in="shadowSpreadOuter1" result="shadowOffsetOuter1" />
-            <feGaussianBlur stdDeviation="6" in="shadowOffsetOuter1" result="shadowBlurOuter1" />
-            <feColorMatrix
-              values="0 0 0 0 0   0 0 0 0 0   0 0 0 0 0  0 0 0 0.1 0"
-              type="matrix"
-              in="shadowBlurOuter1"
-            />
-          </filter>
-          <filter x="-14.0%" y="-10.4%" width="128.1%" height="129.9%" filterUnits="objectBoundingBox">
-            <feGaussianBlur stdDeviation="22" in="SourceAlpha" result="shadowBlurInner1" />
-            <feOffset dx="0" dy="2" in="shadowBlurInner1" result="shadowOffsetInner1" />
-            <feComposite
-              in="shadowOffsetInner1"
-              in2="SourceAlpha"
-              operator="arithmetic"
-              k2="-1"
-              k3="1"
-              result="shadowInnerInner1"
-            />
-            <feColorMatrix
-              values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.493234828 0"
-              type="matrix"
-              in="shadowInnerInner1"
-              result="shadowMatrixInner1"
-            />
-            <feOffset dx="0" dy="-3" in="SourceAlpha" result="shadowOffsetInner2" />
-            <feComposite
-              in="shadowOffsetInner2"
-              in2="SourceAlpha"
-              operator="arithmetic"
-              k2="-1"
-              k3="1"
-              result="shadowInnerInner2"
-            />
-            <feColorMatrix
-              values="0 0 0 0 0.0980392157   0 0 0 0 0.109803922   0 0 0 0 0.137254902  0 0 0 0.472061821 0"
-              type="matrix"
-              in="shadowInnerInner2"
-              result="shadowMatrixInner2"
-            />
-            <feMerge>
-              <feMergeNode in="shadowMatrixInner1" />
-              <feMergeNode in="shadowMatrixInner2" />
-            </feMerge>
-          </filter>
-        </defs>
-        <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-          <g transform="translate(-722.000000, -1858.000000)">
-            <g transform="translate(0.000000, 1628.000000)">
-              <g transform="translate(730.000000, 230.000000)">
-                <g mask="url(#mask-2)">
-                  <g transform="translate(18.016304, 48.644022)">
-                    <path
-                      d="M105.620918,68.6445576 C106.292985,68.205139 106.991123,67.8220542 107.712584,67.499296 L108.662554,61.7670356 C110.606489,61.9131967 111.563209,60.1069852 112.041576,57.7885856 C112.198073,57.0298208 113.239026,52.8890046 112.977753,51.6031262 C112.957248,51.5009848 112.870016,51.4277362 112.768551,51.4277362 C112.667297,51.427591 112.539345,51.5462552 112.510429,51.650247 L110.766833,58.3178936 C110.745624,58.4208335 110.685308,58.5110693 110.599556,58.5681291 C110.513945,58.6256971 110.40952,58.6455882 110.309393,58.6238096 L110.287268,58.6188732 C110.079053,58.5732108 109.946161,58.3623218 109.99013,58.1478031 L111.777713,51.4556854 C111.82598,51.2224371 111.681603,50.9933993 111.455208,50.9438168 C111.338875,50.9184085 111.225502,50.9400418 111.132281,51.0024011 C111.03913,51.0646878 110.974023,51.1626912 110.950841,51.2745604 L109.163257,57.9666055 C109.141766,58.0698358 109.081944,58.1597087 108.996051,58.2172041 C108.910439,58.274772 108.806015,58.2947357 108.705888,58.2727394 C108.475829,58.2221405 108.342656,58.0112515 108.387118,57.7967328 L110.174208,51.1045425 C110.197319,50.9926007 110.176181,50.8757225 110.115865,50.7798243 C110.055197,50.6837809 109.960074,50.6164851 109.851633,50.5925287 C109.62094,50.5422203 109.394545,50.6937264 109.345715,50.9311126 L107.560034,57.6154626 C107.538754,57.7186204 107.47858,57.8088562 107.392827,57.8660612 C107.306934,57.9234114 107.202439,57.9433024 107.102665,57.9215239 L107.08054,57.9165874 C106.872324,57.8711428 106.739433,57.6600361 106.783542,57.4453721 L108.579933,50.7099151 C108.600438,50.6111131 108.552805,50.510714 108.464939,50.4662131 C108.377002,50.4217123 108.270605,50.444362 108.207118,50.5213854 C107.389195,51.5166653 106.060319,55.5031607 105.884164,56.3568074 C105.405938,58.675207 105.799046,60.7700861 107.392475,61.4887057 L105.620918,68.6445576 Z"
-                      stroke="none"
-                      fill="#9FC6E2"
-                      fillRule="evenodd"
-                      transform="translate(109.320266, 59.544082) rotate(22.000000) translate(-109.320266, -59.544082) "
-                    />
-                    <path
-                      d="M74.3819807,44.652395 C76.2269167,45.755483 78.2821099,46.3070269 80.5475604,46.3070269 C82.8130109,46.3070269 84.8682041,45.755483 86.7131401,44.652395"
-                      stroke="#215484"
-                      strokeWidth="1.14087559"
-                      fill="none"
-                      strokeLinecap="round"
-                    />
-                    <path
-                      d="M98.2222222,44.652395 C100.067158,45.755483 102.122351,46.3070269 104.387802,46.3070269 C106.653252,46.3070269 108.708446,45.755483 110.553382,44.652395"
-                      stroke="#215484"
-                      strokeWidth="1.14087559"
-                      fill="none"
-                      strokeLinecap="round"
-                    />
-                    <path
-                      d="M68.6274396,62.8533459 C75.6381964,66.1626097 83.4479307,67.8172416 92.0566425,67.8172416 C100.665354,67.8172416 108.475089,66.1626097 115.485845,62.8533459"
-                      stroke="#215484"
-                      strokeWidth="1.14087559"
-                      fill="none"
-                      strokeLinecap="round"
-                    />
-                    <ellipse
-                      stroke="none"
-                      fill="#215484"
-                      fillRule="evenodd"
-                      cx="92.4676812"
-                      cy="54.1665285"
-                      rx="6.57661836"
-                      ry="1.24097392"
-                    />
-                    <path
-                      d="M68.8722673,32.3510353 C64.9128868,28.2871031 62.4889873,23.4268344 61.6005688,17.7702292 C60.6591416,11.7761143 61.5054214,6.12092111 64.1394084,0.804649758 C66.3259026,7.99130917 68.3089111,12.996868 70.088434,15.8213263 C71.7104679,18.3958182 74.4091044,21.0629707 78.1843433,23.8227838 L78.1843507,23.8227736 C78.8452425,24.3059053 79.005251,25.2254282 78.5464072,25.9034079 C77.685863,27.1749334 76.5553721,28.4243555 75.1549345,29.6516741 C73.6211518,30.995854 72.0961604,32.0051069 70.5799604,32.6794327 L70.579952,32.6794137 C69.9981031,32.9381894 69.3166468,32.8071491 68.8722673,32.3510353 Z"
-                      stroke="none"
-                      fill="#9FC6E2"
-                      fillRule="evenodd"
-                    />
-                    <path
-                      d="M112.953607,32.3510353 C108.994227,28.2871031 106.570327,23.4268344 105.681909,17.7702292 C104.740482,11.7761143 105.586762,6.12092111 108.220748,0.804649758 C110.407243,7.99130917 112.390251,12.996868 114.169774,15.8213263 C115.791808,18.3958182 118.490444,21.0629707 122.265683,23.8227838 L122.265691,23.8227736 C122.926583,24.3059053 123.086591,25.2254282 122.627747,25.9034079 C121.767203,27.1749334 120.636712,28.4243555 119.236275,29.6516741 C117.702492,30.995854 116.1775,32.0051069 114.6613,32.6794327 L114.661292,32.6794137 C114.079443,32.9381894 113.397987,32.8071491 112.953607,32.3510353 Z"
-                      stroke="none"
-                      fill="#9FC6E2"
-                      fillRule="evenodd"
-                      transform="translate(114.097313, 16.808765) scale(-1, 1) translate(-114.097313, -16.808765) "
-                    />
-                    <g
-                      stroke="none"
-                      strokeWidth="1"
-                      fill="none"
-                      fillRule="evenodd"
-                      transform="translate(0.000000, 0.260661)"
-                    >
-                      <ellipse fill="#FFFAC2" cx="92.4836957" cy="99.2439146" rx="92.4836957" ry="9.92779139" />
-                      <g transform="translate(16.014493, 0.543989)">
-                        <g transform="translate(0.000000, 13.237055)">
-                          <path
-                            d="M1.81639595,91.3220226 L5.47948376,84.2867994 L0.49909202,90.2921764 C2.7812653,82.4584252 4.59389657,77.0994211 5.93698582,74.2151642 C7.28007508,71.3309073 9.56927363,67.673546 12.8045815,63.2430803 L3.84362246,71.238176 C5.81213828,67.1218999 7.45459671,64.077679 8.77099774,62.1055134 C10.0873988,60.1333478 13.1094366,56.4292945 17.8371112,50.9933535 L7.18218796,57.696358 C10.4774026,50.7508681 13.5470483,45.7124895 16.3911251,42.5812222 C19.235202,39.4499548 24.3330508,35.57849 31.6846716,30.9668276 C29.4643846,31.2375494 27.7122995,31.5553909 26.4284162,31.920352 C25.1445329,32.2853131 23.290424,33.103371 20.8660893,34.3745258 C26.0217141,28.2386083 29.9845802,24.0814198 32.7546877,21.9029603 C36.909849,18.6352711 40.8099243,15.7780663 50.808187,11.7141609 C57.4736955,9.00489063 63.9353012,7.35853791 70.1930041,6.77510277 L67.2642142,1.50573774 C69.1237082,1.98505962 70.6580763,2.61435612 71.8673187,3.39362722 C73.0765611,4.17289833 74.1373718,5.30005684 75.0497507,6.77510277 L77.4499828,0.326149012 L77.4499828,6.25762503 C79.6919814,4.53566228 81.4533799,3.30104628 82.7341783,2.55377702 C84.0149767,1.80650775 85.2220612,1.30919148 86.3554319,1.0618282 C84.9112078,2.29442564 83.9460599,3.26208089 83.4599881,3.96479397 C82.9739162,4.66750705 82.6217594,5.60427664 82.4035175,6.77510277 C93.0009851,8.34389388 101.386368,11.3182213 107.559667,15.6980851 C113.732966,20.0779489 121.197958,28.4429274 129.954643,40.7930207 L121.419147,35.969662 C126.612272,40.532819 130.445936,44.4160764 132.920138,47.6194344 C135.394341,50.8227924 137.592347,54.181767 139.514159,57.696358 L129.350627,50.9933535 C134.857972,56.7852098 138.68044,61.2614742 140.81803,64.4221467 C142.95562,67.5828193 145.295564,72.1175477 147.837862,78.0263321 L139.514159,72.1919222 C144.744988,78.3097318 148.062767,82.4501116 149.467497,84.6130617 C150.872227,86.7760118 151.698735,88.3576491 151.947022,89.3579738 L147.36633,86.52033 L151.094748,91.3220226 C131.852921,94.0446928 106.504588,95.406028 75.0497507,95.406028 C43.5090342,95.406028 19.097916,94.0446928 1.81639595,91.3220226 Z"
-                            fill="#FFFFFF"
-                          />
-                          <path
-                            d="M45.5075469,67.6465878 C45.0978056,65.3507496 44.849435,61.950435 44.7600264,57.4377021 C44.7550137,57.1846922 44.9550546,56.9755036 45.2068303,56.9704662 C45.458606,56.9654289 45.6667742,57.1664504 45.671787,57.4194603 C45.7603227,61.8881402 46.0055854,65.2459054 46.4051664,67.484814 C46.802221,69.7095674 47.5566286,71.9749861 48.6696114,74.2814102 L49.9395428,76.9130776 L42.8650332,69.4678761 L48.6728204,85.4424814 L43.0687105,79.7104706 L50.3600689,91.9160037 L46.3661118,90.562805 L49.0251406,94.9925118 C47.9040633,94.9590437 47.9099824,94.9590437 47.9040633,94.9590437 C47.9001173,94.9590437 46.6372467,92.9170819 44.1154513,88.8331585 L48.302186,90.2516724 L38.3261504,73.5520637 L46.3703093,81.7798121 L40.511012,65.6635257 L46.8189772,72.3020172 C46.2190055,70.7347863 45.7817423,69.1829431 45.5075469,67.6465878 Z"
-                            fill="#215484"
-                            fillRule="nonzero"
-                          />
-                          <path
-                            d="M107.964069,67.6465878 C107.554327,65.3507496 107.305957,61.950435 107.216548,57.4377021 C107.211535,57.1846922 107.411576,56.9755036 107.663352,56.9704662 C107.915128,56.9654289 108.123296,57.1664504 108.128309,57.4194603 C108.216844,61.8881402 108.462107,65.2459054 108.861688,67.484814 C109.258743,69.7095674 110.01315,71.9749861 111.126133,74.2814102 L112.396064,76.9130776 L105.321555,69.4678761 L111.129342,85.4424814 L105.525232,79.7104706 L112.816591,91.9160037 L108.822634,90.562805 L111.481662,94.9925118 C110.360585,94.9590437 110.366504,94.9590437 110.360585,94.9590437 C110.356639,94.9590437 109.093768,92.9170819 106.571973,88.8331585 L110.758708,90.2516724 L100.782672,73.5520637 L108.826831,81.7798121 L102.967534,65.6635257 L109.275499,72.3020172 C108.675527,70.7347863 108.238264,69.1829431 107.964069,67.6465878 Z"
-                            fill="#215484"
-                            fillRule="nonzero"
-                            transform="translate(106.799631, 75.981443) scale(-1, 1) translate(-106.799631, -75.981443) "
-                          />
-                        </g>
-                        <path
-                          d="M89.6064251,67.8399078 C90.2784923,67.4004892 90.9766305,67.0174044 91.6980916,66.6946463 L92.6480612,60.9623859 C94.5919961,61.1085469 95.5487167,59.3023355 96.0270837,56.9839359 C96.18358,56.2251711 97.2245332,52.0843548 96.9632598,50.7984764 C96.9427554,50.696335 96.8555233,50.6230865 96.7540579,50.6230865 C96.652804,50.6229413 96.5248526,50.7416054 96.4959359,50.8455973 L94.7523403,57.5132438 C94.7311313,57.6161838 94.6708157,57.7064196 94.5850634,57.7634794 C94.499452,57.8210474 94.3950272,57.8409384 94.2949006,57.8191599 L94.2727755,57.8142234 C94.0645601,57.768561 93.9316686,57.5576721 93.975637,57.3431533 L95.7632204,50.6510356 C95.8114869,50.4177873 95.6671101,50.1887495 95.4407155,50.139167 C95.3243826,50.1137587 95.2110091,50.1353921 95.1177878,50.1977513 C95.024637,50.260038 94.95953,50.3580415 94.936348,50.4699107 L93.1487646,57.1619557 C93.1272737,57.2651861 93.0674514,57.3550589 92.9815581,57.4125543 C92.8959467,57.4701223 92.7915219,57.4900859 92.6913953,57.4680896 C92.4613367,57.4174908 92.3281634,57.2066018 92.3726249,56.992083 L94.1597151,50.2998927 C94.1828267,50.187951 94.161688,50.0710727 94.1013725,49.9751745 C94.0407047,49.8791311 93.9455809,49.8118354 93.8371398,49.787879 C93.606447,49.7375705 93.3800523,49.8890766 93.3312221,50.1264629 L91.5455412,56.8108129 C91.5242617,56.9139706 91.4640871,57.0042064 91.3783347,57.0614114 C91.2924415,57.1187616 91.1879462,57.1386527 91.0881719,57.1168741 L91.0660468,57.1119377 C90.8578314,57.0664931 90.72494,56.8553863 90.7690492,56.6407224 L92.5654404,49.9052654 C92.5859448,49.8064633 92.5383125,49.7060642 92.4504463,49.6615633 C92.3625096,49.6170625 92.2561119,49.6397122 92.1926256,49.7167357 C91.3747019,50.7120156 90.0458266,54.698511 89.8696714,55.5521577 C89.3914454,57.8705573 89.7845532,59.9654363 91.3779824,60.684056 L89.6064251,67.8399078 Z"
-                          fill="#9FC6E2"
-                          transform="translate(93.305773, 58.739432) rotate(22.000000) translate(-93.305773, -58.739432) "
-                        />
-                        <path
-                          d="M58.3674879,43.8477453 C60.2124239,44.9508332 62.2676172,45.5023772 64.5330676,45.5023772 C66.7985181,45.5023772 68.8537113,44.9508332 70.6986473,43.8477453"
-                          stroke="#215484"
-                          strokeWidth="1.14087559"
-                          strokeLinecap="round"
-                        />
-                        <path
-                          d="M82.2077295,43.8477453 C84.0526655,44.9508332 86.1078587,45.5023772 88.3733092,45.5023772 C90.6387597,45.5023772 92.6939529,44.9508332 94.5388889,43.8477453"
-                          stroke="#215484"
-                          strokeWidth="1.14087559"
-                          strokeLinecap="round"
-                        />
-                        <path
-                          d="M52.6129469,62.0486962 C59.6237037,65.35796 67.433438,67.0125919 76.0421498,67.0125919 C84.6508616,67.0125919 92.4605959,65.35796 99.4713527,62.0486962"
-                          stroke="#215484"
-                          strokeWidth="1.14087559"
-                          strokeLinecap="round"
-                        />
-                        <ellipse
-                          fill="#215484"
-                          cx="76.4531884"
-                          cy="53.3618787"
-                          rx="6.57661836"
-                          ry="1.24097392"
-                        />
-                        <path
-                          d="M52.8577746,31.5463855 C48.898394,27.4824533 46.4744945,22.6221846 45.5860761,16.9655794 C44.6446488,10.9714645 45.4909287,5.31627136 48.1249156,0 C50.3114098,7.18665941 52.2944183,12.1922183 54.0739412,15.0166766 C55.6959751,17.5911685 58.3946116,20.258321 62.1698506,23.018134 L62.169858,23.0181239 C62.8307497,23.5012556 62.9907582,24.4207785 62.5319145,25.0987582 C61.6713702,26.3702837 60.5408793,27.6197057 59.1404418,28.8470243 C57.606659,30.1912042 56.0816677,31.2004571 54.5654676,31.8747829 L54.5654592,31.874764 C53.9836103,32.1335397 53.3021541,32.0024993 52.8577746,31.5463855 Z"
-                          fill="#9FC6E2"
-                        />
-                        <path
-                          d="M96.9391147,31.5463855 C92.9797341,27.4824533 90.5558346,22.6221846 89.6674161,16.9655794 C88.7259889,10.9714645 89.5722688,5.31627136 92.2062557,0 C94.3927499,7.18665941 96.3757584,12.1922183 98.1552813,15.0166766 C99.7773152,17.5911685 102.475952,20.258321 106.251191,23.018134 L106.251198,23.0181239 C106.91209,23.5012556 107.072098,24.4207785 106.613255,25.0987582 C105.75271,26.3702837 104.622219,27.6197057 103.221782,28.8470243 C101.687999,30.1912042 100.163008,31.2004571 98.6468077,31.8747829 L98.6467993,31.874764 C98.0649504,32.1335397 97.3834942,32.0024993 96.9391147,31.5463855 Z"
-                          fill="#9FC6E2"
-                          transform="translate(98.082820, 16.004116) scale(-1, 1) translate(-98.082820, -16.004116) "
-                        />
-                      </g>
-                    </g>
-                  </g>
-                </g>
-              </g>
-            </g>
-          </g>
-        </g>
-      </svg>
-    );
-  }
-}
-
-export class ExploreButtonIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="14px" height="8px" viewBox="0 0 14 8">
-        <path
-          fill="#FFFFFF"
-          d="M6.88461538,0.00037425682 C7.12873677,-0.00644425667 7.37571617,0.0798702667 7.57156599,0.262257984 L13.4423916,5.72954163 C13.8476876,6.1069788 13.8805433,6.7529224 13.515777,7.17229703 C13.1510106,7.59167167 12.5267518,7.6256687 12.1214559,7.24823153 L6.88461538,2.37135491 L1.64777489,7.24823153 C1.24247896,7.6256687 0.618220141,7.59167167 0.253453799,7.17229703 C-0.111312542,6.7529224 -0.0784568152,6.1069788 0.32683912,5.72954163 L6.19766478,0.262257984 C6.39351459,0.0798702667 6.640494,-0.00644425667 6.88461538,0.00037425682 Z"
-          transform="translate(6.884615, 3.755245) scale(1, -1) translate(-6.884615, -3.755245) "
-        />
-      </svg>
-    )
-  }
-}
-
 export class RightArrowIcon extends PureComponent {
   static propTypes = {
     color: PropTypes.string
@@ -537,26 +52,6 @@ export class RightArrowIcon extends PureComponent {
         <path
           d="M7.51011525,6.88461538 C7.51693377,7.12873677 7.43061924,7.37571617 7.24823153,7.57156599 L1.78094788,13.4423916 C1.40351071,13.8476876 0.757567115,13.8805433 0.338192479,13.515777 C-0.0811821572,13.1510106 -0.115179189,12.5267518 0.262257984,12.1214559 L5.1391346,6.88461538 L0.262257984,1.64777489 C-0.115179189,1.24247896 -0.0811821572,0.618220141 0.338192479,0.253453799 C0.757567115,-0.111312542 1.40351071,-0.0784568152 1.78094788,0.32683912 L7.24823153,6.19766478 C7.43061924,6.39351459 7.51693377,6.640494 7.51011525,6.88461538 Z"
         />
-      </svg>
-    )
-  }
-}
-
-export class LetsGoHomeButtonIcon extends PureComponent {
-  render() {
-    return (
-      <svg width="14px" height="6px">
-        <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-          <g transform="translate(-285.000000, -453.000000)" fill="#FFFFFF">
-            <g transform="translate(127.000000, 197.000000)">
-              <g transform="translate(0.000000, 234.000000)">
-                <g transform="translate(141.000000, 0.000000)">
-                  <path d="M23.8846154,22.0002746 C24.1287368,21.9952718 24.3757162,22.0586013 24.571566,22.1924202 L30.4423916,26.2037978 C30.8476876,26.4807256 30.8805433,26.9546581 30.515777,27.2623557 C30.1510106,27.5700533 29.5267518,27.5949971 29.1214559,27.3180693 L23.8846154,23.7398768 L18.6477749,27.3180693 C18.242479,27.5949971 17.6182201,27.5700533 17.2534538,27.2623557 C16.8886875,26.9546581 16.9215432,26.4807256 17.3268391,26.2037978 L23.1976648,22.1924202 C23.3935146,22.0586013 23.640494,21.9952718 23.8846154,22.0002746 Z" transform="translate(23.884615, 24.755245) scale(1, -1) translate(-23.884615, -24.755245) " />
-                </g>
-              </g>
-            </g>
-          </g>
-        </g>
       </svg>
     )
   }
