+++ src/pages/LPMA/sections/Header/icons/index.js
@@ -1,0 +1,52 @@
+/* eslint-disable */
+import React, {PureComponent} from 'react';
+
+export class LPMAGraph extends PureComponent {
+  render() {
+    return (
+      <svg viewBox="0 0 1508 391">
+        <g fill="none" fillRule="evenodd">
+          <g transform="translate(0 -4)">
+            <g fontFamily="Rubik-Regular, Rubik" fontSize="20.16">
+              <text fill="#8C9BA2" transform="translate(53.76 371.28)">
+                <tspan x="0" y="19">Apr 02, 2018</tspan>
+              </text>
+              <text fill="#8C9BA2" transform="translate(53.76 371.28)">
+                <tspan x="332.64" y="19">Apr 22, 2018</tspan>
+              </text>
+              <text fill="#3D474C" transform="translate(53.76 371.28)">
+                <tspan x="668.64" y="19">Apr 29, 2018</tspan>
+              </text>
+              <text fill="#8C9BA2" transform="translate(53.76 371.28)">
+                <tspan x="1004.64" y="19">Jun 14, 2018</tspan>
+              </text>
+              <text fill="#8C9BA2" transform="translate(53.76 371.28)">
+                <tspan x="1333.92" y="19">Jun 25, 2018</tspan>
+              </text>
+            </g>
+            <path d="M53.76 120.7343c19.7566 9.2015 201.0462 121.1658 411.5932 121.1658 284.838 0 398.9187-183.1001 638.6188-183.1001 239.7 0 309.5805 103.8795 401.308 140.6807" stroke="#6EBF56" strokeWidth="4.86" />
+            <path d="M53.76 120.167c19.7566 8.9545 201.0462 117.9131 411.5932 117.9131 284.838 0 398.9187-176.0394 638.6188-176.0394 239.7 0 304.932 85.6597 401.308 134.7587V330.96H53.76V120.167z" fillOpacity=".04" fill="#49D14E" />
+            <g strokeLinecap="square" strokeWidth="1.68">
+              <path d="M54.6 331.8h1449.84" stroke="#D7D7D7" />
+              <path d="M53.76 170.52h1451.52M53.76 10.92h1451.52" strokeOpacity=".1022" stroke="#738187" />
+            </g>
+            <circle stroke="#6CBF55" strokeWidth="3.36" fill="#FFF" cx="1500.24" cy="198.24" r="5.04" />
+            <circle stroke="#6CBF55" strokeWidth="3.36" fill="#FFF" cx="450.24" cy="241.92" r="5.04" />
+            <circle stroke="#6CBF55" strokeWidth="3.36" fill="#FFF" cx="58.8" cy="122.64" r="5.04" />
+            <g fill="#8C9BA2" fontFamily="Rubik-Regular, Rubik" fontSize="20.16">
+              <text>
+                <tspan x="0" y="339.88">0</tspan>
+              </text>
+              <text>
+                <tspan x="0" y="178.6">5</tspan>
+              </text>
+              <text>
+                <tspan x="0" y="19">10</tspan>
+              </text>
+            </g>
+          </g>
+        </g>
+      </svg>
+    );
+  }
+}
