+++ src/pages/Landing/sections/Services/icons/index.js
@@ -1,0 +1,180 @@
+/* eslint-disable */
+import React, {PureComponent} from 'react';
+
+export class ApisIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="62px" height="49px" viewBox="0 0 62 49">
+        <g fill="none" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" stroke="#0084FF">
+          <path d="M52.2,28.7 C57.0599743,27.4871863 60.2728874,22.8657731 59.7169604,17.8876999 C59.1610335,12.9096266 55.0077924,9.11082963 50,9 C47.7136247,3.40930845 41.8777497,0.12987513 35.9134657,1.08416056 C29.9491817,2.038446 25.4278693,6.97502932 25,13" />
+          <path d="M54,35 C54,28.372583 48.627417,23 42,23 L40.8,23 C38.0648659,16.2555713 31.1225782,12.2067816 23.905648,13.1470982 C16.6887178,14.0874149 11.0156514,19.7798989 10.1,27 L10,27 C4.4771525,27 4.4408921e-16,31.4771525 0,37 C-4.4408921e-16,42.5228475 4.4771525,47 10,47 L42.1,47 C48.6883314,46.9450953 54.0002288,41.5885602 54,35 Z" />
+        </g>
+      </svg>
+    );
+  }
+}
+
+export class AutomationIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="62px" height="48px" viewBox="0 0 62 48">
+        <g fill="none" stroke="#39C877" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+          <path d="M52.741,20.14 C53.9623053,29.8912495 48.8536697,39.3413102 40.0261325,43.660227 C31.1985954,47.9791438 20.6022818,46.2128083 13.653,39.264 M7.065,25.62 C6.3207602,19.1066924 8.39031651,12.5852041 12.754065,7.69288682 C17.1178134,2.80056951 23.3613105,0.00206701696 29.917,-3.65614318e-15 C36.0172875,-0.00820191744 41.8694582,2.41442834 46.179,6.732" />
+          <path d="M0,16 L7.065,25.619 L16,18 M60,30 L52.741,20.14 L44,28" />
+          <circle cx="30" cy="22.994" r="6" />
+        </g>
+      </svg>
+    );
+  }
+}
+
+export class BusinessAnalysisIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="48px" height="48px" viewBox="0 0 48 48">
+        <g fill="none" stroke="#39C877" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+          <path d="M19.9333333,6.13333333 C8.92445732,6.13333333 4.91458726e-15,15.0577907 3.55271368e-15,26.0666667 C2.1908401e-15,37.0755427 8.92445732,46 19.9333333,46 C30.9422093,46 39.8666667,37.0755427 39.8666667,26.0666667 L19.9333333,26.0666667 L19.9333333,6.13333333 Z" />
+          <path d="M26.0666667,-1.77635684e-15 L26.0666667,19.9333333 L46,19.9333333 C46,14.6466846 43.899887,9.57656331 40.1616618,5.83833816 C36.4234367,2.10011302 31.3533154,-1.43588845e-15 26.0666667,-1.77635684e-15 Z" />
+        </g>
+      </svg>
+    );
+  }
+}
+
+export class DesignThinkingIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="50px" height="47px" viewBox="0 0 50 47">
+        <path fill="#FFC600" d="M35.0363768,22.7809069 C35.3915827,22.3580036 36.0223649,22.3031241 36.4452681,22.65833 C36.8681714,23.013536 36.923051,23.6443181 36.5678451,24.0672214 C34.4206468,26.623646 31.4700576,28.3587883 28.2085328,28.9822173 C27.6660692,29.0859074 27.1422581,28.7302111 27.0385681,28.1877474 C26.9348781,27.6452837 27.2905744,27.1214727 27.833038,27.0177827 C30.6419702,26.4808651 33.1845615,24.9856527 35.0363768,22.7809069 Z M39.473836,18.5517344 C40.9850864,18.9818977 42.4489939,19.6658937 43.8031263,20.6118262 C49.9352508,24.895437 51.8019581,33.1434854 48.1123593,39.6516834 C44.4226334,46.1601056 36.3881354,48.7908308 29.5657029,45.7242332 C29.0619659,45.4978098 28.8371589,44.9058979 29.0635823,44.402161 C29.2900057,43.8984241 29.8819176,43.6736171 30.3856546,43.9000405 C36.266374,46.5433499 43.19187,44.2757437 46.3725007,38.6653305 C49.5532585,33.054693 47.9439767,25.9440796 42.657795,22.2514042 C37.5022117,18.6499588 30.5007369,19.4533699 26.2903015,24.0598606 C27.2162537,25.3529658 27.9418461,26.8180086 28.4116458,28.4200592 C29.7240335,32.8953947 28.7791986,37.534531 26.1982772,41.0646686 C26.2577811,41.4027599 26.1414824,41.7625859 25.8610034,42.0009523 C25.6806382,42.1542366 25.4616527,42.2324164 25.2421254,42.2385825 C23.7949014,43.8369754 21.9710427,45.1330977 19.8439146,45.9766131 C12.881989,48.7373737 4.96308766,45.7410416 1.57254372,39.0631416 C-1.8180005,32.385241 0.435877412,24.2239335 6.7729413,20.2324649 C7.24025406,19.9381229 7.85769726,20.0783434 8.15203929,20.5456561 C8.44638131,21.0129689 8.30616083,21.6304121 7.83884807,21.9247541 C2.37599158,25.3655937 0.433039776,32.4010355 3.35585296,38.157709 C6.27866642,43.914383 13.1051449,46.4973671 19.1066638,44.117457 C20.9184282,43.3989995 22.4749736,42.2996468 23.7152969,40.9444032 C21.0919906,37.2857456 20.3042325,32.5885893 21.6381634,28.241 C21.8001622,27.7130086 22.3595097,27.416313 22.887501,27.5783117 C23.4154923,27.7403105 23.712188,28.2996579 23.5501892,28.8276493 C22.460192,32.3802022 23.0134631,36.2035043 24.999881,39.2784813 C26.90689,36.3211742 27.5503388,32.5902854 26.4924641,28.9828573 C24.8790801,23.4811028 19.7657139,19.8535692 14.1752942,20.0044166 C15.0414657,21.8834145 16.3743627,23.5686074 18.1337206,24.8673565 C18.5780534,25.1953608 18.6723564,25.8214637 18.3443521,26.2657965 C18.0163479,26.7101293 17.390245,26.8044323 16.9459122,26.476428 C14.6493106,24.7810884 12.9804107,22.5190388 12.005439,20.0086854 C11.7557786,19.8605475 11.5723765,19.6054613 11.5261871,19.2959535 C11.4988849,19.1130066 11.5229859,18.9342484 11.5876421,18.7743322 C10.5060489,15.04596 10.8933791,10.9071515 12.9489732,7.31630129 C16.6615633,0.830898858 24.6211246,-1.77996561 31.3976123,1.2689743 C37.9787118,4.23000335 41.3794981,11.6004627 39.473836,18.5517344 Z M37.512783,18.1383352 C39.22919,12.0963834 36.2898144,5.66323217 30.5769903,3.09286561 C24.7416163,0.467360268 17.8870908,2.71575518 14.6846928,8.30991971 C12.9402158,11.3572896 12.586049,14.8619251 13.4579896,18.0382233 C17.8673644,17.7161087 22.0775899,19.4331086 25.0045019,22.5060608 C28.3018679,19.0335953 32.9960142,17.5095764 37.512783,18.1383352 Z" />
+      </svg>
+    );
+  }
+}
+
+export class DesktopAppsIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="62px" height="44px" viewBox="0 0 62 44">
+        <g fill="none" stroke="#0084FF" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+          <path d="M4,37 L4,2 C4,0.8954305 4.8954305,0 6,0 L54,0 C55.1045695,0 56,0.8954305 56,2 L56,37" />
+          <path d="M0,37 L0,40 C0,41.1045695 0.8954305,42 2,42 L58,42 C59.1045695,42 60,41.1045695 60,40 L60,37 L0,37 Z" />
+        </g>
+      </svg>
+    );
+  }
+}
+
+export class ContactUsIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="28px" height="28px" viewBox="0 0 28 28">
+        <path fill="#005EFF" d="M14,0 C6.28064877,0 -2.84217094e-14,6.28064877 -2.84217094e-14,14 C-2.84217094e-14,21.7193512 6.28064877,28 14,28 C21.7193512,28 28,21.7193512 28,14 C28,6.28064877 21.7193512,0 14,0 Z" />
+        <path fill="#FFFFFF" d="M9,8 C9,11.2 9,12.8 9,12.8 C9,12.8 11,13.2 14.9999999,14 L9,15.2 L9,20 L21,14 L9,8 Z" />
+      </svg>
+    );
+  }
+}
+
+export class MobileAppsIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="32px" height="50px" viewBox="0 0 32 50">
+        <rect transform="translate(1, 1)" fill="none" stroke="#0084FF" strokeWidth="2" strokeLinecap="round"
+              strokeLinejoin="round" x="0" y="0" width="30" height="48" rx="4.34782609"
+        />
+        <circle transform="translate(1, 1)" fill="#0084FF" cx="15" cy="41" r="2" />
+      </svg>
+    );
+  }
+}
+
+export class QAIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="50px" height="50px" viewBox="0 0 50 50">
+        <g fill="none" stroke="#39C877" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+          <circle cx="24" cy="24" r="24" />
+          <polyline points="34.4 17.28 22.0923077 31.28 14.4 23.5022222" />
+        </g>
+      </svg>
+    );
+  }
+}
+
+export class PrototypeIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="58px" height="54px" viewBox="0 0 58 54">
+        <g transform="translate(1.000000, 1.000000)" fill="none" stroke="#FFC600">
+          <rect strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" x="0" y="0" width="56" height="52" rx="6" />
+          <g strokeWidth="2.08" transform="translate(14.000000, 9.000000)">
+            <path strokeLinecap="square" d="M7.60630005,30.9107524 C21.7854128,30.9107524 7.1878154,6.00339627 20.7440961,6.00339627" />
+            <ellipse cx="3.73333333" cy="31.3970588" rx="3.73333333" ry="3.60294118" />
+            <polygon transform="translate(22.082821, 5.974843) rotate(-226.000000) translate(-22.082821, -5.974843) " points="26.3494873 2.61446454 17.816154 1.85719612 18.5067128 10.0924902" />
+          </g>
+        </g>
+      </svg>
+    );
+  }
+}
+
+export class ProjectManagementIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="48px" height="48px" viewBox="0 0 48 48">
+        <g fill="none" stroke="#39C877" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+          <path d="M21.0833333,42.1666667 C9.43932986,42.1666667 1.36187358e-15,32.7273368 0,21.0833333 C-1.70234197e-15,9.43932986 9.43932986,2.04281037e-15 21.0833333,0 C32.7273368,-2.21304456e-15 42.1666667,9.43932986 42.1666667,21.0833333 M21.0833333,33.35 C14.3086404,33.35 8.81666667,27.8580263 8.81666667,21.0833333 C8.81666667,14.3086404 14.3086404,8.81666667 21.0833333,8.81666667 C27.8580263,8.81666667 33.35,14.3086404 33.35,21.0833333" />
+          <path d="M46,46 L27.5233333,27.5233333 M40.7866667,28.75 L27.5233333,27.5233333 L28.5966667,40.8633333" />
+        </g>
+      </svg>
+    );
+  }
+}
+
+export class UIDesignIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="58px" height="51px" viewBox="0 0 58 51">
+        <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd" strokeLinecap="round" strokeLinejoin="round">
+          <g transform="translate(1.000000, 1.000000)" stroke="#FFC600">
+            <path strokeWidth="2" d="M6,0 L50,0 C53.3137085,-6.08718376e-16 56,2.6862915 56,6 L56,11.3076923 L0,11.3076923 L0,6 C-4.05812251e-16,2.6862915 2.6862915,6.08718376e-16 6,0 Z M0,11.3076923 L0,43 C4.05812251e-16,46.3137085 2.6862915,49 6,49 L50,49 C53.3137085,49 56,46.3137085 56,43 L56,11.3076923 L0,11.3076923 Z" />
+            <ellipse id="Oval" strokeWidth="1.86666667" cx="5.6" cy="5.65384615" rx="1" ry="1" />
+            <ellipse id="Oval" strokeWidth="1.86666667" cx="12.1333333" cy="5.65384615" rx="1" ry="1" />
+            <ellipse id="Oval" strokeWidth="1.86666667" cx="18.6666667" cy="5.65384615" rx="1" ry="1" />
+          </g>
+        </g>
+      </svg>
+
+    );
+  }
+}
+
+export class UXDesignIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="54px" height="55px" viewBox="0 0 54 55">
+        <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd" strokeLinecap="round" strokeLinejoin="round">
+          <g id="vector-1546_55acc497-0fcb-45c8-828f-c40e7536b7e9" transform="translate(1.000000, 1.000000)" stroke="#FFC600" strokeWidth="1.79310345">
+            <path d="M0,0 L8.66666667,0 L8.66666667,8.83333333 L0,8.83333333 L0,0 Z M43.3333333,0 L52,0 L52,8.83333333 L43.3333333,8.83333333 L43.3333333,0 Z M0,44.1666667 L8.66666667,44.1666667 L8.66666667,53 L0,53 L0,44.1666667 Z M43.3333333,44.1666667 L52,44.1666667 L52,53 L43.3333333,53 L43.3333333,44.1666667 Z" />
+            <path d="M43.3333333,48.5833333 L8.66666667,48.5833333 M47.6666667,8.83333333 L47.6666667,44.1666667 M8.66666667,4.41666667 L43.3333333,4.41666667 M4.33333333,44.1666667 L4.33333333,8.83333333" />
+          </g>
+        </g>
+      </svg>
+    );
+  }
+}
+
+export class WebAppsIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="56px" height="47px" viewBox="0 0 56 47">
+        <g fill="none" stroke="#0084FF" transform="translate(1, 1)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+          <path strokeWidth="2" d="M5.54658841,0 L48.4534116,0 C51.5167078,-5.62718381e-16 54,2.48329222 54,5.54658841 L54,12.8571429 L0,12.8571429 L8.8817842e-16,5.54658841 C5.13032832e-16,2.48329222 2.48329222,5.62718381e-16 5.54658841,0 Z M0,12.8571429 L0,39.4534116 C3.75145588e-16,42.5167078 2.48329222,45 5.54658841,45 L48.4534116,45 C51.5167078,45 54,42.5167078 54,39.4534116 L54,12.8571429 L0,12.8571429 Z" />
+          <ellipse strokeWidth="0.924431401" fill="#0084FF" cx="6.47762289" cy="6.5224359" rx="1" ry="1" />
+          <ellipse strokeWidth="0.924431401" fill="#0084FF" cx="12.9486427" cy="6.5224359" rx="1" ry="1" />
+          <ellipse strokeWidth="0.924431401" fill="#0084FF" cx="19.4196625" cy="6.5224359" rx="1" ry="1" />
+        </g>
+      </svg>
+    );
+  }
+}
