+++ src/pages/Page404/icons/index.js
@@ -1,0 +1,22 @@
+/* eslint-disable */
+import React, {PureComponent} from 'react';
+
+export class LetsGoHomeButtonIcon extends PureComponent {
+  render() {
+    return (
+      <svg width="14px" height="6px">
+        <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+          <g transform="translate(-285.000000, -453.000000)" fill="#FFFFFF">
+            <g transform="translate(127.000000, 197.000000)">
+              <g transform="translate(0.000000, 234.000000)">
+                <g transform="translate(141.000000, 0.000000)">
+                  <path d="M23.8846154,22.0002746 C24.1287368,21.9952718 24.3757162,22.0586013 24.571566,22.1924202 L30.4423916,26.2037978 C30.8476876,26.4807256 30.8805433,26.9546581 30.515777,27.2623557 C30.1510106,27.5700533 29.5267518,27.5949971 29.1214559,27.3180693 L23.8846154,23.7398768 L18.6477749,27.3180693 C18.242479,27.5949971 17.6182201,27.5700533 17.2534538,27.2623557 C16.8886875,26.9546581 16.9215432,26.4807256 17.3268391,26.2037978 L23.1976648,22.1924202 C23.3935146,22.0586013 23.640494,21.9952718 23.8846154,22.0002746 Z" transform="translate(23.884615, 24.755245) scale(1, -1) translate(-23.884615, -24.755245) " />
+                </g>
+              </g>
+            </g>
+          </g>
+        </g>
+      </svg>
+    )
+  }
+}
