+++ src/pages/Page404/index.js
@@ -2,15 +2,18 @@ import React, {PureComponent} from 'react';
 import Heading from 'components/Typography/Heading';
 import Paragraph from 'components/Typography/Paragraph';
 import Button from 'components/Button';
-import {LetsGoHomeButtonIcon} from 'icons';
+import Loader from 'components/Loader';
+import {darkBlue, blue} from 'styles/config';
 
-import './index.styl';
 import Illustration from './Illustration';
+import {LetsGoHomeButtonIcon} from './icons';
+import './index.styl';
 
 class Page404 extends PureComponent {
   render() {
     return (
       <div className="page-404">
+        <Loader colors={{light: blue, dark: darkBlue}} />
         <div className="content">
           <div className="content-info">
             <Heading subTitle="404" title="Page not found" />
